refactor(serve): split serve.rs into a module directory
serve.rs had grown to roughly 5,000 lines, the largest file in the repository,
having absorbed every dashboard feature: 30-odd routes, SSE streaming, the
wizard and estimate flow, review and corrections, validation, glossary, the
whole audiobook surface, runtime leases, child spawning with scrubbed
environments, and the inlined frontend assets. Three separate changes landed in
it this week alone. commands/translate/ was split for the same reason.

serve.rs is now 502 lines — router assembly and shared state — with domain
modules alongside it: assets, security, jobs, translation, glossary, audio,
options, and tests.

This is a pure move. No behaviour change, no route change, no renamed field, no
altered header. Every visibility addition is `pub(super)` and exists only
because a parent or sibling module boundary requires it; no public or
crate-visible API changed. All 226 previously named items are preserved, with
six private route-assembly functions added.

Two invariants were held deliberately:

- The test count is exactly 846, unchanged. A pure move adds and removes no
  tests, so any other number would mean something was lost or duplicated.
- `dashboard_assets_reassemble_byte_stably` passes with its original
  113,292-byte length and unchanged SHA-256, and the three frontend assets have
  an empty diff. That test exists to make asset changes deliberate; it is the
  mechanism that would have caught the double-encoded separator that shipped
  earlier this week, so passing it without editing the constant is the proof
  that the shipped dashboard is byte-identical.

One pre-existing non-fatal cargo path-canonicalization warning was left
untouched, and no bug found during the move was fixed here — a refactor that
also changes behaviour cannot be reviewed.

846 passed, 0 failed, 3 ignored. clippy -D warnings and fmt clean.
