+++ crates/bookforge-cli/src/commands/serve.rs
@@ -47,6 +47,22 @@ use serde_json::json;
 
 use crate::eventlog::{EventLogTailer, events_path_for};
 
+mod assets;
+mod audio;
+mod glossary;
+mod jobs;
+mod options;
+mod security;
+mod translation;
+
+use super::{audiobook, correct, estimate, reconfigure, review, validate};
+use security::{
+    generate_csrf_token, reject_mutation, require_loopback_bind, validate_dashboard_host,
+};
+use translation::{
+    configure_dashboard_child_environment, provider_key_env, resolve_dashboard_provider_key,
+};
+
 /// Where browser-launched uploads and their outputs are written.
 const UPLOAD_DIR: &str = ".bookforge/serve-uploads";
 
@@ -357,104 +373,18 @@ pub async fn run(args: ServeArgs) -> Result<()> {
 fn dashboard_router(state: AppState) -> Router {
     let host_state = state.clone();
     Router::new()
-        .route("/", get(index))
-        .route("/api/jobs", get(list_jobs))
-        .route("/api/jobs/{id}", get(job_detail))
-        .route(
-            "/api/jobs/{id}/reconfigure",
-            get(job_reconfigure).post(update_job_reconfigure),
-        )
-        .route("/api/jobs/{id}/events", get(job_events))
-        .route("/api/jobs/{id}/review", get(job_review))
-        .route(
-            "/api/jobs/{id}/segments/{segment_id}/translation",
-            post(save_manual_translation),
-        )
-        .route(
-            "/api/jobs/{id}/segments/{segment_id}/flag",
-            post(set_segment_flag),
-        )
-        .route(
-            "/api/jobs/{id}/segments/{segment_id}/retry",
-            post(retry_segment_with_guidance),
-        )
-        .route("/api/jobs/{id}/validate", post(job_validate))
-        .route("/api/jobs/{id}/retry", post(retry_job))
-        .route("/api/jobs/{id}/pause", post(pause_job))
-        .route("/api/jobs/{id}/resume", post(resume_job))
-        .route("/api/jobs/{id}/stop", post(stop_job))
-        .route("/api/options", get(dashboard_options))
-        .route("/api/providers", get(provider_status))
-        .route("/api/audio/voices", get(audio_voices))
-        .route("/api/estimate", post(estimate_translate))
-        .route("/api/translate", post(launch_translate))
-        .route("/api/audiobook/estimate", post(estimate_audiobook))
-        .route("/api/audiobook", post(launch_audiobook))
-        .route("/api/audiobooks", get(list_audiobooks))
-        .route("/api/audiobooks/{id}", get(audiobook_status))
-        .route("/api/audiobooks/{id}/cancel", post(cancel_audiobook))
-        .route("/api/audiobooks/{id}/artifact", get(audiobook_artifact))
-        .route("/api/glossary", get(list_glossary).post(add_glossary))
-        .route("/api/glossary/{id}", delete(remove_glossary))
+        .merge(assets::routes())
+        .merge(jobs::routes())
+        .merge(options::routes())
+        .merge(audio::routes())
+        .merge(translation::routes())
+        .merge(glossary::routes())
         .layer(DefaultBodyLimit::max(MAX_UPLOAD_BYTES))
         .layer(middleware::from_fn_with_state(
             host_state,
             validate_dashboard_host,
         ))
         .with_state(state)
-}
-
-async fn validate_dashboard_host(
-    State(state): State<AppState>,
-    request: Request,
-    next: Next,
-) -> Response {
-    if dashboard_host_allowed(request.headers(), state.host_port) {
-        return next.run(request).await;
-    }
-    forbidden("dashboard host header rejected")
-}
-
-fn dashboard_host_allowed(headers: &HeaderMap, port: u16) -> bool {
-    headers
-        .get(HOST)
-        .and_then(|value| value.to_str().ok())
-        .is_some_and(|host| dashboard_host_value_allowed(host, port))
-}
-
-fn dashboard_host_value_allowed(host: &str, port: u16) -> bool {
-    if host.is_empty()
-        || host
-            .bytes()
-            .any(|byte| matches!(byte, b' ' | b'\t' | b'\r' | b'\n'))
-    {
-        return false;
-    }
-
-    if let Some(rest) = host.strip_prefix('[') {
-        let Some((addr, suffix)) = rest.split_once(']') else {
-            return false;
-        };
-        return addr == "::1" && suffix == format!(":{port}");
-    }
-
-    let Some((name, host_port)) = host.rsplit_once(':') else {
-        return false;
-    };
-    if name.contains(':') || host_port != port.to_string() {
-        return false;
-    }
-
-    name == "127.0.0.1" || name.eq_ignore_ascii_case("localhost")
-}
-
-fn require_loopback_bind(addr: SocketAddr) -> Result<()> {
-    if addr.ip().is_loopback() {
-        return Ok(());
-    }
-    anyhow::bail!(
-        "--bind must use a loopback address such as 127.0.0.1:8765; use an SSH tunnel for remote access"
-    );
 }
 
 async fn shutdown_signal() {
@@ -464,1901 +394,6 @@ async fn shutdown_signal() {
 // ---------------------------------------------------------------------------
 // Handlers
 // ---------------------------------------------------------------------------
-
-async fn index(State(state): State<AppState>) -> Response {
-    (
-        [
-            ("content-security-policy", DASHBOARD_CONTENT_SECURITY_POLICY),
-            ("x-frame-options", "DENY"),
-            ("x-content-type-options", "nosniff"),
-            ("referrer-policy", "no-referrer"),
-            ("cache-control", "no-store"),
-        ],
-        Html(DASHBOARD_HTML.replace(CSRF_TOKEN_PLACEHOLDER, &state.csrf_token)),
-    )
-        .into_response()
-}
-
-async fn list_jobs(State(state): State<AppState>) -> Result<Json<Vec<JobListItem>>, AppError> {
-    let store_path = state.store_path.clone();
-    let items = tokio::task::spawn_blocking(move || -> Result<Vec<JobListItem>> {
-        let store = JobStore::open(store_path)?;
-        Ok(store
-            .list_job_summaries()?
-            .into_iter()
-            .map(|(job, summary)| JobListItem::new(&job, &summary))
-            .collect())
-    })
-    .await??;
-    Ok(Json(items))
-}
-
-async fn job_detail(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-) -> Result<Response, AppError> {
-    let lookup = id.clone();
-    let store_path = state.store_path.clone();
-    let detail = tokio::task::spawn_blocking(move || -> Result<Option<JobDetail>> {
-        let store = JobStore::open(store_path)?;
-        let job = store.get_job(&lookup)?;
-        let events_path = events_path_for(job.as_ref(), &lookup);
-        if job.is_none() && !events_path.exists() {
-            return Ok(None);
-        }
-        let mut tailer = EventLogTailer::new(events_path);
-        let mut state = RunState::default();
-        for event in tailer.poll()? {
-            state.fold(&event);
-        }
-        Ok(Some(JobDetail::new(lookup, job, state)))
-    })
-    .await??;
-
-    match detail {
-        Some(detail) => Ok(Json(detail).into_response()),
-        None => Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({ "error": format!("no job '{id}'") })),
-        )
-            .into_response()),
-    }
-}
-
-async fn job_reconfigure(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-) -> Result<Response, AppError> {
-    let store_path = state.store_path.clone();
-    let lease_stale_after = state.runtime_lease_stale_after;
-    let outcome = tokio::task::spawn_blocking(move || {
-        runtime_settings_view(&store_path, &id, lease_stale_after)
-    })
-    .await?;
-    match outcome {
-        Ok(Some(view)) => Ok(Json(view).into_response()),
-        Ok(None) => Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({ "error": "no such job or run snapshot" })),
-        )
-            .into_response()),
-        Err(error) => Ok(bad_request(&error.to_string())),
-    }
-}
-
-async fn update_job_reconfigure(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Json(incoming): Json<super::reconfigure::RunConfigOverrides>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    if incoming.is_empty() {
-        return Ok(bad_request("select at least one runtime setting"));
-    }
-    let store_path = state.store_path.clone();
-    let lease_stale_after = state.runtime_lease_stale_after;
-    let outcome = tokio::task::spawn_blocking(move || -> Result<RuntimeSettingsView> {
-        let store = JobStore::open(&store_path)?;
-        let Some(job) = store.get_job(&id)? else {
-            anyhow::bail!("no such job");
-        };
-        if !matches!(job.status.as_str(), "running" | "paused" | "stopped") {
-            anyhow::bail!(
-                "job '{}' is {}; runtime settings are editable only while running, paused, or stopped",
-                id,
-                job.status
-            );
-        }
-        if store.load_job_config_snapshot(&id)?.is_none() {
-            anyhow::bail!("job '{}' has no resumable run snapshot", id);
-        }
-        let (_path, written) =
-            super::reconfigure::write_merged_overrides_for_job(&id, incoming)?;
-        let mut view = runtime_settings_view(&store_path, &id, lease_stale_after)?
-            .ok_or_else(|| anyhow::anyhow!("job disappeared after reconfiguration"))?;
-        view.revision = written.revision;
-        Ok(view)
-    })
-    .await?;
-
-    match outcome {
-        Ok(view) => Ok(Json(view).into_response()),
-        Err(error) => Ok(bad_request(&error.to_string())),
-    }
-}
-
-async fn job_events(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-) -> Sse<impl futures_core::Stream<Item = Result<Event, Infallible>>> {
-    let refresh = state.refresh;
-    let path = resolve_events_path(id, state.store_path.clone()).await;
-
-    let stream = async_stream::stream! {
-        let mut tailer = EventLogTailer::new(path);
-        let mut run = RunState::default();
-        let mut ticker = tokio::time::interval(refresh);
-        ticker.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Skip);
-        let mut last = String::new();
-
-        loop {
-            ticker.tick().await;
-            if let Ok(events) = tailer.poll() {
-                for event in events {
-                    run.fold(&event);
-                }
-            }
-            if let Ok(payload) = serde_json::to_string(&run)
-                && payload != last
-            {
-                last = payload.clone();
-                yield Ok(Event::default().event("state").data(payload));
-            }
-            if run.finished {
-                yield Ok(Event::default().event("done").data("done"));
-                break;
-            }
-        }
-    };
-
-    Sse::new(stream).keep_alive(KeepAlive::default())
-}
-
-/// Serve a job's side-by-side review as JSON. Shares
-/// [`generate_review_document`](super::review::generate_review_document) with the
-/// CLI `review` command; the browser renders it into the Review screen. Errors
-/// (unknown job, or a job that predates run-config snapshots) become 404s.
-async fn job_review(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-) -> Result<Response, AppError> {
-    let store_path = state.store_path.clone();
-    let outcome = tokio::task::spawn_blocking(move || {
-        let store = JobStore::open(store_path)?;
-        super::review::generate_review_document(&store, &id)
-    })
-    .await?;
-
-    match outcome {
-        Ok(document) => Ok(Json(document).into_response()),
-        Err(err) => Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({ "error": err.to_string() })),
-        )
-            .into_response()),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct ManualCorrectionRequest {
-    blocks: Vec<super::correct::CorrectionBlock>,
-}
-
-async fn save_manual_translation(
-    AxumPath((id, segment_id)): AxumPath<(String, String)>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Json(request): Json<ManualCorrectionRequest>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    if request.blocks.is_empty() {
-        return Ok(bad_request("at least one corrected block is required"));
-    }
-
-    let store_path = state.store_path.clone();
-    // Held across the whole read-modify-write so a second tab cannot stage a
-    // rebuilt EPUB from a snapshot taken before this correction lands.
-    let lock = job_correction_lock(&state, &id)?;
-    let _guard = lock.lock().await;
-    let outcome = tokio::task::spawn_blocking(move || -> Result<_> {
-        let store = JobStore::open(store_path)?;
-        super::correct::correct_job_segment(
-            &store,
-            &id,
-            &segment_id,
-            super::correct::CorrectionPayload::Blocks(request.blocks),
-        )
-    })
-    .await?;
-
-    match outcome {
-        Ok(outcome) => Ok(Json(outcome).into_response()),
-        Err(err) => Ok(bad_request(&err.to_string())),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct SegmentFlagRequest {
-    flagged: bool,
-}
-
-async fn set_segment_flag(
-    AxumPath((id, segment_id)): AxumPath<(String, String)>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Json(request): Json<SegmentFlagRequest>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    let store_path = state.store_path.clone();
-    let outcome = tokio::task::spawn_blocking(move || -> Result<()> {
-        let store = JobStore::open(store_path)?;
-        store.set_dashboard_segment_flag(&id, &segment_id, request.flagged)?;
-        Ok(())
-    })
-    .await?;
-    match outcome {
-        Ok(()) => Ok(Json(json!({ "flagged": request.flagged })).into_response()),
-        Err(err) => Ok(bad_request(&err.to_string())),
-    }
-}
-
-#[derive(Debug, Deserialize)]
-struct SegmentRetryRequest {
-    #[serde(default)]
-    guidance: Option<String>,
-}
-
-async fn retry_segment_with_guidance(
-    AxumPath((id, segment_id)): AxumPath<(String, String)>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Json(request): Json<SegmentRetryRequest>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    let store_path = state.store_path.clone();
-    let outcome = tokio::task::spawn_blocking(move || -> Result<()> {
-        let store = JobStore::open(store_path)?;
-        store.request_segment_retry(&id, &segment_id, request.guidance.as_deref())?;
-        Ok(())
-    })
-    .await?;
-    match outcome {
-        Ok(()) => Ok(Json(json!({ "retry_pending": true })).into_response()),
-        Err(err) => Ok(bad_request(&err.to_string())),
-    }
-}
-
-/// Validate a job's translated EPUB (BookForge structural validators + EPUBCheck)
-/// and return the report. Shares [`validate_path`](super::validate::validate_path)
-/// with the CLI `validate` command. EPUBCheck may be `unavailable` (no external
-/// tool) — that's reported, not an error. POST + CSRF because it runs an external
-/// process and is triggered explicitly ("Re-run check").
-async fn job_validate(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-
-    let store_path = state.store_path.clone();
-    let outcome = tokio::task::spawn_blocking(
-        move || -> Result<Option<super::validate::ValidationReport>> {
-            let store = JobStore::open(store_path)?;
-            let Some(job) = store.get_job(&id)? else {
-                return Ok(None);
-            };
-            if !job.output_path.exists() {
-                anyhow::bail!(
-                    "translated EPUB not found at {} — finish the run first",
-                    job.output_path.display()
-                );
-            }
-            Ok(Some(
-                super::validate::validate_path(&job.output_path, false).report,
-            ))
-        },
-    )
-    .await?;
-
-    match outcome {
-        Ok(Some(report)) => Ok(Json(report).into_response()),
-        Ok(None) => Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({ "error": "no such job" })),
-        )
-            .into_response()),
-        Err(err) => Ok(bad_request(&err.to_string())),
-    }
-}
-
-async fn retry_job(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    let store_path = state.store_path.clone();
-    let retried = tokio::task::spawn_blocking(move || -> Result<usize> {
-        let store = JobStore::open(store_path)?;
-        Ok(store.retry_segments(&id, RetryScope::All)?)
-    })
-    .await??;
-    Ok(Json(json!({ "retried": retried })).into_response())
-}
-
-async fn pause_job(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Response, AppError> {
-    control_job(id, state, headers, ControlCommand::Pause).await
-}
-
-async fn resume_job(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    request: Option<Json<ResumeJobRequest>>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    let request = request.map(|Json(request)| request).unwrap_or_default();
-    let store_path = state.store_path.clone();
-    let lease_stale_after = state.runtime_lease_stale_after;
-    let lookup = id.clone();
-    let action = tokio::task::spawn_blocking(move || -> Result<Option<ResumeJobAction>> {
-        let store = JobStore::open(&store_path)?;
-        let Some(job) = store.get_job(&lookup)? else {
-            return Ok(None);
-        };
-        let snapshot = store.load_job_config_snapshot(&lookup)?;
-        let live = matches!(
-            crate::control::runtime_lease_state(&lookup, lease_stale_after),
-            crate::control::RuntimeLeaseState::Fresh(_)
-        );
-        let resumable = !store.resumable_segment_ids(&lookup)?.is_empty()
-            || (job_status_has_unfinished_pipeline_work(&job.status) && snapshot.is_some());
-        let force = !live && job.status == "paused";
-        Ok(Some(ResumeJobAction {
-            live,
-            resumable,
-            force,
-            provider: job.provider,
-            api_key_env: snapshot.and_then(|snapshot| snapshot.api_key_env),
-        }))
-    })
-    .await??;
-    let Some(action) = action else {
-        return Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({ "error": "no such job" })),
-        )
-            .into_response());
-    };
-    if action.live {
-        let path = crate::control::request_job_control(&id, ControlCommand::Resume)?;
-        return Ok(Json(json!({
-            "command": "resume",
-            "mode": "signaled",
-            "control_path": path,
-        }))
-        .into_response());
-    }
-    if !action.resumable {
-        return Ok(bad_request(
-            "the worker is not alive and this job has no resumable work",
-        ));
-    }
-
-    let api_key_env = if action.provider == "mock" {
-        None
-    } else {
-        action
-            .api_key_env
-            .or_else(|| provider_key_env(&action.provider).map(str::to_string))
-    };
-    let key = resolve_dashboard_provider_key(
-        &state,
-        &action.provider,
-        request.api_key,
-        api_key_env.as_deref(),
-    )?;
-    if action.provider != "mock" && key.is_none() {
-        return Ok(missing_resume_key(&action.provider, api_key_env.as_deref()));
-    }
-
-    let Some(mut launch_claim) = crate::control::RuntimeLaunchClaim::acquire(&id)? else {
-        return Ok(Json(json!({
-            "command": "resume",
-            "mode": "launching",
-        }))
-        .into_response());
-    };
-    let executable =
-        std::env::current_exe().context("failed to locate the BookForge executable")?;
-    let mut command = tokio::process::Command::new(executable);
-    command
-        .arg("resume")
-        .arg(&id)
-        .arg("--ui")
-        .arg("quiet")
-        .stdin(std::process::Stdio::null())
-        .stdout(std::process::Stdio::null())
-        .stderr(std::process::Stdio::null());
-    if action.force {
-        // A paused job normally expects to signal its original process. A
-        // missing/stale lease proves that process is unavailable, so the
-        // replacement must use the CLI's explicit dead-worker escape hatch.
-        command.arg("--force");
-    }
-    #[cfg(windows)]
-    {
-        use std::os::windows::process::CommandExt;
-        command.as_std_mut().creation_flags(0x0800_0000);
-    }
-    configure_dashboard_child_environment(&mut command, api_key_env.as_deref().zip(key.as_deref()));
-    #[cfg(test)]
-    if let Some(launches) = &state.resume_launches {
-        launches.fetch_add(1, std::sync::atomic::Ordering::SeqCst);
-        if let Some(environments) = &state.resume_child_environments {
-            let environment = command
-                .as_std()
-                .get_envs()
-                .map(|(name, value)| (name.to_os_string(), value.map(ToOwned::to_owned)))
-                .collect();
-            environments
-                .lock()
-                .map_err(|_| anyhow::anyhow!("resume environment recorder is unavailable"))?
-                .push(environment);
-        }
-        launch_claim.persist_until_worker();
-        return Ok(Json(json!({
-            "command": "resume",
-            "mode": "spawned",
-            "pid": 0,
-            "forced": action.force,
-        }))
-        .into_response());
-    }
-    let mut child = command.spawn().context("failed to launch resume worker")?;
-    let pid = child.id();
-    if let Some(status) = child_exit_status_after(&mut child, CHILD_STARTUP_CHECK).await?
-        && !status.success()
-    {
-        return Ok(bad_request(&format!(
-            "resume worker exited immediately with {status}"
-        )));
-    }
-    launch_claim.persist_until_worker();
-    Ok(Json(json!({
-        "command": "resume",
-        "mode": "spawned",
-        "pid": pid,
-        "forced": action.force,
-    }))
-    .into_response())
-}
-
-#[derive(Default, Deserialize)]
-struct ResumeJobRequest {
-    api_key: Option<String>,
-}
-
-struct ResumeJobAction {
-    live: bool,
-    resumable: bool,
-    force: bool,
-    provider: String,
-    api_key_env: Option<String>,
-}
-
-fn missing_resume_key(provider: &str, api_key_env: Option<&str>) -> Response {
-    let env_note = api_key_env
-        .map(|env| format!(" ({env})"))
-        .unwrap_or_default();
-    (
-        StatusCode::BAD_REQUEST,
-        Json(json!({
-            "error": format!(
-                "API key for provider '{provider}'{env_note} is unavailable; supply it to resume this job"
-            ),
-            "requires_api_key": true,
-            "provider": provider,
-            "api_key_env": api_key_env,
-        })),
-    )
-        .into_response()
-}
-
-fn job_status_has_unfinished_pipeline_work(status: &str) -> bool {
-    matches!(status, "running" | "paused" | "stopped")
-}
-
-async fn stop_job(
-    AxumPath(id): AxumPath<String>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Response, AppError> {
-    control_job(id, state, headers, ControlCommand::Stop).await
-}
-
-async fn control_job(
-    id: String,
-    state: AppState,
-    headers: HeaderMap,
-    command: ControlCommand,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    let store_path = state.store_path.clone();
-    let lease_stale_after = state.runtime_lease_stale_after;
-    let outcome = tokio::task::spawn_blocking(move || -> Result<Option<String>> {
-        let store = JobStore::open(store_path)?;
-        if store.get_job(&id)?.is_none() {
-            return Ok(None);
-        }
-        if !matches!(
-            crate::control::runtime_lease_state(&id, lease_stale_after),
-            crate::control::RuntimeLeaseState::Fresh(_)
-        ) {
-            anyhow::bail!(
-                "no live worker is available for {}; refresh the job and use Resume to launch one",
-                command.as_str()
-            );
-        }
-        let path = crate::control::request_job_control(&id, command)?;
-        Ok(Some(path.display().to_string()))
-    })
-    .await?;
-
-    match outcome {
-        Ok(Some(path)) => Ok(Json(json!({
-            "command": command.as_str(),
-            "control_path": path,
-        }))
-        .into_response()),
-        Ok(None) => Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({ "error": "no such job" })),
-        )
-            .into_response()),
-        Err(error) => Ok(bad_request(&error.to_string())),
-    }
-}
-
-async fn dashboard_options() -> Json<DashboardOptions> {
-    Json(dashboard_options_payload())
-}
-
-/// Launch a new translation from an uploaded EPUB.
-///
-/// Runs the translation as a detached `bookforge translate` subprocess. The
-/// child receives an allowlisted subset of this process's environment and may
-/// receive a dashboard-supplied session key through the provider's normal key env
-/// var; key values are never placed on the command line. The job is matched back
-/// to the dashboard by its unique input path (returned to the client), since the
-/// run generates its own job id.
-async fn launch_translate(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    mut multipart: Multipart,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-
-    let mut file_bytes: Option<Vec<u8>> = None;
-    let mut file_name = "upload.epub".to_string();
-    let mut fields: HashMap<String, String> = HashMap::new();
-
-    while let Some(field) = multipart.next_field().await? {
-        let name = field.name().unwrap_or_default().to_string();
-        if name == "file" {
-            if let Some(fname) = field.file_name()
-                && !fname.is_empty()
-            {
-                file_name = fname.to_string();
-            }
-            file_bytes = Some(field.bytes().await?.to_vec());
-        } else {
-            fields.insert(name, field.text().await?);
-        }
-    }
-
-    let Some(bytes) = file_bytes.filter(|b| !b.is_empty()) else {
-        return Ok(bad_request("upload an EPUB file"));
-    };
-    let Some(target) = field_value(&fields, "target") else {
-        return Ok(bad_request("target language is required"));
-    };
-    let provider = field_value(&fields, "provider").unwrap_or_else(|| "mock".to_string());
-    if !supported_provider(&provider) {
-        return Ok(bad_request("unsupported provider"));
-    }
-    let openai_base_url = if provider == "openai-compatible" {
-        let Some(base_url) = field_value(&fields, "base_url") else {
-            return Ok(bad_request("base URL is required for openai-compatible"));
-        };
-        if !dashboard_base_url_uses_https(&base_url) {
-            return Ok(bad_request(
-                "base URL must use https:// for openai-compatible",
-            ));
-        }
-        if field_value(&fields, "model").is_none() {
-            return Ok(bad_request("model is required for openai-compatible"));
-        }
-        Some(base_url)
-    } else {
-        None
-    };
-
-    let supplied_key = (provider != "mock")
-        .then(|| field_value(&fields, "api_key"))
-        .flatten();
-    let key = resolve_dashboard_provider_key(
-        &state,
-        &provider,
-        supplied_key,
-        provider_key_env(&provider),
-    )?;
-    if provider != "mock" && key.is_none() {
-        return Ok(bad_request("provider API key is required"));
-    }
-
-    let stem = sanitize_component(strip_epub_suffix(&file_name));
-    let tag = format!("{}-{stem}", now_ms());
-    let upload_dir = state.upload_dir.clone();
-    std::fs::create_dir_all(&upload_dir)?;
-    let input_path = upload_dir.join(format!("{tag}.epub"));
-    std::fs::write(&input_path, &bytes)?;
-    let out_path = upload_dir.join(format!("{tag}.{}.epub", sanitize_component(&target)));
-
-    let exe = std::env::current_exe()?;
-    let mut command = tokio::process::Command::new(exe);
-    command
-        .arg("translate")
-        .arg(&input_path)
-        .arg("--target")
-        .arg(&target)
-        .arg("--provider")
-        .arg(&provider)
-        .arg("--ui")
-        .arg("quiet")
-        .arg("--out")
-        .arg(&out_path);
-    if let Some(source) = field_value(&fields, "source") {
-        command.arg("--source").arg(source);
-    }
-    // Offline mock runs are identity translations unless told otherwise.
-    let model = field_value(&fields, "model")
-        .or_else(|| (provider == "mock").then(|| "mock-identity".to_string()));
-    if let Some(model) = model {
-        command.arg("--model").arg(model);
-    }
-    if let Some(profile) = field_value(&fields, "profile") {
-        command.arg("--profile").arg(profile);
-    }
-    if provider == "deepseek" {
-        command.arg("--no-thinking");
-    }
-    // Advanced tuning from the wizard, each validated before forwarding so the
-    // child never receives arbitrary argv from the browser.
-    if let Some(concurrency) = field_value(&fields, "concurrency")
-        .and_then(|value| value.parse::<usize>().ok())
-        .map(|value| value.clamp(1, 16))
-    {
-        command.arg("--concurrency").arg(concurrency.to_string());
-    }
-    if let Some(context) = field_value(&fields, "context_window")
-        .and_then(|value| value.parse::<usize>().ok())
-        .map(|value| value.min(16))
-    {
-        command.arg("--context-window").arg(context.to_string());
-    }
-    if let Some(qa) = field_value(&fields, "qa")
-        .filter(|value| matches!(value.as_str(), "off" | "suspicious" | "all"))
-    {
-        command.arg("--qa").arg(qa);
-    }
-    if field_value(&fields, "validate_output")
-        .is_some_and(|value| matches!(value.as_str(), "true" | "on" | "1"))
-    {
-        command.arg("--validate-output");
-    }
-    if let Some(base_url) = openai_base_url {
-        command.arg("--base-url").arg(base_url);
-    }
-    let api_key_env = (provider != "mock" && key.is_some())
-        .then(|| provider_key_env(&provider).expect("provider was validated"));
-    configure_dashboard_child_environment(&mut command, api_key_env.zip(key.as_deref()));
-
-    // Point the run at the one canonical provider key copied into its otherwise
-    // scrubbed environment. The child records the env-var name in its job
-    // snapshot, so `bookforge resume` can use the same name later.
-    if let Some(env) = api_key_env {
-        command.arg("--api-key-env").arg(env);
-    }
-
-    // Detached: the run outlives this request. The short startup check catches
-    // immediate argv/binary failures before the dashboard reports success.
-    let mut child = command
-        .spawn()
-        .context("failed to spawn translation process")?;
-    let pid = child.id();
-    let completed_immediately = if let Some(status) =
-        child_exit_status_after(&mut child, CHILD_STARTUP_CHECK).await?
-    {
-        if !status.success() {
-            return Err(anyhow::anyhow!(
-                    "translation process exited immediately with {status}; check the serve console for details"
-                )
-                .into());
-        }
-        true
-    } else {
-        false
-    };
-
-    Ok(Json(json!({
-        "ok": true,
-        "input_path": input_path.display().to_string(),
-        "provider": provider,
-        "pid": pid,
-        "completed_immediately": completed_immediately,
-    }))
-    .into_response())
-}
-
-struct AudiobookSource {
-    bytes: Vec<u8>,
-    file_name: String,
-}
-
-/// Resolve either a direct browser upload or the output of a finished
-/// translation. The latter is addressed by job id rather than by a
-/// browser-supplied filesystem path, keeping the handoff scoped to records the
-/// server already trusts.
-async fn resolve_audiobook_source(
-    state: &AppState,
-    fields: &HashMap<String, String>,
-    file_bytes: Option<Vec<u8>>,
-    file_name: String,
-) -> Result<AudiobookSource> {
-    if let Some(bytes) = file_bytes.filter(|bytes| !bytes.is_empty()) {
-        return Ok(AudiobookSource { bytes, file_name });
-    }
-
-    let source_job_id = field_value(fields, "source_job_id")
-        .context("upload an EPUB file or choose a finished translation")?;
-    let store_path = state.store_path.clone();
-    tokio::task::spawn_blocking(move || -> Result<AudiobookSource> {
-        let store = JobStore::open(store_path)?;
-        let job = store
-            .get_job(&source_job_id)?
-            .with_context(|| format!("no translation job '{source_job_id}'"))?;
-        if job.status != "succeeded" {
-            anyhow::bail!("translation must be finished before it can be narrated");
-        }
-        let metadata = std::fs::metadata(&job.output_path).with_context(|| {
-            format!(
-                "translated EPUB is not available at {}",
-                job.output_path.display()
-            )
-        })?;
-        if metadata.len() > MAX_UPLOAD_BYTES as u64 {
-            anyhow::bail!("translated EPUB exceeds the dashboard upload limit");
-        }
-        let file_name = job
-            .output_path
-            .file_name()
-            .and_then(|name| name.to_str())
-            .unwrap_or("translated.epub")
-            .to_string();
-        let bytes = std::fs::read(&job.output_path).with_context(|| {
-            format!(
-                "failed to read translated EPUB at {}",
-                job.output_path.display()
-            )
-        })?;
-        if bytes.is_empty() {
-            anyhow::bail!("translated EPUB is empty");
-        }
-        Ok(AudiobookSource { bytes, file_name })
-    })
-    .await?
-}
-
-/// Plan audiobook synthesis from an uploaded EPUB or finished translation
-/// without making provider requests. Parsing and chunk construction are
-/// blocking, so both happen off the async worker after the source has been
-/// saved to a temporary path.
-async fn estimate_audiobook(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    mut multipart: Multipart,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-
-    let mut file_bytes: Option<Vec<u8>> = None;
-    let mut file_name = "upload.epub".to_string();
-    let mut fields = HashMap::<String, String>::new();
-    while let Some(field) = multipart.next_field().await? {
-        let name = field.name().unwrap_or_default().to_string();
-        if name == "file" {
-            if let Some(value) = field.file_name().filter(|value| !value.is_empty()) {
-                file_name = value.to_string();
-            }
-            file_bytes = Some(field.bytes().await?.to_vec());
-        } else {
-            fields.insert(name, field.text().await?);
-        }
-    }
-    let source = match resolve_audiobook_source(&state, &fields, file_bytes, file_name).await {
-        Ok(source) => source,
-        Err(error) => return Ok(bad_request(&error.to_string())),
-    };
-    let bytes = source.bytes;
-
-    let provider = field_value(&fields, "provider").unwrap_or_else(|| "mock".to_string());
-    if !matches!(
-        provider.as_str(),
-        "mock" | "openai" | "gemini" | "elevenlabs"
-    ) {
-        return Ok(bad_request("unsupported audiobook provider"));
-    }
-    let model = field_value(&fields, "model").unwrap_or_else(|| match provider.as_str() {
-        "openai" => "gpt-4o-mini-tts".to_string(),
-        "gemini" => "gemini-3.1-flash-tts-preview".to_string(),
-        "elevenlabs" => String::new(),
-        _ => "mock-silence".to_string(),
-    });
-    let max_chars = match fields.get("max_chars") {
-        Some(value) => match value.trim().parse::<usize>() {
-            Ok(value) => value,
-            Err(_) => return Ok(bad_request("max_chars must be a positive integer")),
-        },
-        None => 2_000,
-    };
-    let provider_max_chars = audio_provider_max_chars(&provider, &model);
-    if max_chars == 0 || max_chars > provider_max_chars {
-        return Ok(bad_request(&format!(
-            "{provider} accepts between 1 and {provider_max_chars} characters per request"
-        )));
-    }
-    let chapter_filter = match field_value(&fields, "chapters") {
-        Some(value) => match super::audiobook::parse_chapter_ranges(&value) {
-            Ok(chapters) => Some(chapters),
-            Err(error) => return Ok(bad_request(&format!("invalid chapters: {error}"))),
-        },
-        None => None,
-    };
-
-    let sequence = ESTIMATE_SEQ.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    let temp_path = std::env::temp_dir().join(format!(
-        "bookforge-audio-estimate-{}-{}-{sequence}.epub",
-        std::process::id(),
-        now_ms(),
-    ));
-    std::fs::write(&temp_path, bytes)?;
-    let plan_path = temp_path.clone();
-    let plan_result = tokio::task::spawn_blocking(move || -> Result<(usize, usize, usize)> {
-        let book = bookforge_epub::read_epub(&plan_path)?;
-        let options = bookforge_audio::AudiobookOptions {
-            max_chars,
-            chapter_filter,
-            ..bookforge_audio::AudiobookOptions::default()
-        };
-        let plan = bookforge_audio::plan_chunks(&book, &options);
-        let chapters = plan
-            .iter()
-            .map(|chunk| chunk.chapter_index)
-            .collect::<std::collections::BTreeSet<_>>()
-            .len();
-        let characters = plan.iter().map(|chunk| chunk.chars).sum();
-        Ok((chapters, plan.len(), characters))
-    })
-    .await;
-    let _ = std::fs::remove_file(&temp_path);
-    let result = match plan_result? {
-        Ok(result) => result,
-        Err(error) => {
-            return Ok(bad_request(&format!(
-                "could not estimate audiobook: {error}"
-            )));
-        }
-    };
-    let (chapters, chunks, characters) = result;
-    let cost = crate::audio_cost::estimate_audio_cost(&provider, &model, characters);
-    let mut payload = json!({
-        "chapters": chapters,
-        "chunks": chunks,
-        "characters": characters,
-        "est_cost_usd": cost.and_then(|value| value.usd),
-        "est_credits": cost.and_then(|value| value.credits),
-    });
-
-    if provider == "elevenlabs"
-        && let Some(api_key) = resolve_audio_provider_key(&state, "elevenlabs")?
-        && let Some(subscription) = fetch_dashboard_elevenlabs_subscription(&api_key).await
-    {
-        let remaining = subscription
-            .character_limit
-            .saturating_sub(subscription.character_count);
-        if let Some(object) = payload.as_object_mut() {
-            object.insert(
-                "quota".to_string(),
-                json!({
-                    "remaining": remaining,
-                    "limit": subscription.character_limit,
-                    "fits": characters as u128 <= remaining as u128,
-                }),
-            );
-        }
-    }
-
-    Ok(Json(payload).into_response())
-}
-
-async fn fetch_dashboard_elevenlabs_subscription(
-    api_key: &str,
-) -> Option<bookforge_audio::ElevenLabsSubscription> {
-    bookforge_audio::fetch_elevenlabs_subscription_with_key(
-        ELEVENLABS_BASE_URL,
-        api_key,
-        ELEVENLABS_VOICE_TIMEOUT_SECONDS,
-    )
-    .await
-    .ok()
-}
-
-/// Launch audiobook synthesis directly from an uploaded source or translated
-/// EPUB. Progress is durable: the audio builder creates `manifest.json` before
-/// its first provider request and atomically checkpoints it after every chunk.
-async fn launch_audiobook(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    mut multipart: Multipart,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-
-    let mut file_bytes: Option<Vec<u8>> = None;
-    let mut file_name = "upload.epub".to_string();
-    let mut fields = HashMap::<String, String>::new();
-    while let Some(field) = multipart.next_field().await? {
-        let name = field.name().unwrap_or_default().to_string();
-        if name == "file" {
-            if let Some(value) = field.file_name().filter(|value| !value.is_empty()) {
-                file_name = value.to_string();
-            }
-            file_bytes = Some(field.bytes().await?.to_vec());
-        } else {
-            fields.insert(name, field.text().await?);
-        }
-    }
-    let source = match resolve_audiobook_source(&state, &fields, file_bytes, file_name).await {
-        Ok(source) => source,
-        Err(error) => return Ok(bad_request(&error.to_string())),
-    };
-    let bytes = source.bytes;
-    let file_name = source.file_name;
-
-    let provider = field_value(&fields, "provider").unwrap_or_else(|| "mock".to_string());
-    if !matches!(
-        provider.as_str(),
-        "mock" | "openai" | "gemini" | "elevenlabs"
-    ) {
-        return Ok(bad_request("unsupported audiobook provider"));
-    }
-    let model = field_value(&fields, "model").unwrap_or_else(|| match provider.as_str() {
-        "openai" => "gpt-4o-mini-tts".to_string(),
-        "gemini" => "gemini-3.1-flash-tts-preview".to_string(),
-        "elevenlabs" => String::new(),
-        _ => "mock-silence".to_string(),
-    });
-    let auto_model = provider == "elevenlabs" && model.is_empty();
-    let voice = field_value(&fields, "voice").or_else(|| match provider.as_str() {
-        "openai" => Some("alloy".to_string()),
-        "gemini" => Some("Kore".to_string()),
-        "mock" => Some("mock".to_string()),
-        _ => None,
-    });
-    let Some(voice) = voice else {
-        return Ok(bad_request("ElevenLabs requires a voice ID"));
-    };
-    let format = field_value(&fields, "format").unwrap_or_else(|| {
-        if matches!(provider.as_str(), "gemini" | "mock") {
-            "wav".to_string()
-        } else {
-            "mp3".to_string()
-        }
-    });
-    let allowed_format = matches!(
-        format.as_str(),
-        "mp3" | "opus" | "aac" | "flac" | "wav" | "pcm"
-    );
-    if !allowed_format {
-        return Ok(bad_request("unsupported audio format"));
-    }
-    if provider == "mock" && format != "wav" {
-        return Ok(bad_request("the mock provider supports WAV only"));
-    }
-    if provider == "gemini" && !matches!(format.as_str(), "wav" | "pcm") {
-        return Ok(bad_request("Gemini supports WAV or PCM only"));
-    }
-    if provider == "elevenlabs" && matches!(format.as_str(), "aac" | "flac") {
-        return Ok(bad_request("ElevenLabs supports MP3, Opus, WAV, or PCM"));
-    }
-
-    let speed = field_value(&fields, "speed")
-        .and_then(|value| value.parse::<f32>().ok())
-        .unwrap_or(1.0);
-    if !speed.is_finite() || !(0.25..=4.0).contains(&speed) {
-        return Ok(bad_request("speed must be between 0.25 and 4.0"));
-    }
-    if provider == "gemini" && (speed - 1.0).abs() > f32::EPSILON {
-        return Ok(bad_request("Gemini does not expose playback-speed control"));
-    }
-    if provider == "elevenlabs" && model == "eleven_v3" && (speed - 1.0).abs() > f32::EPSILON {
-        return Ok(bad_request(
-            "eleven_v3 does not support speed control; use speed 1.0",
-        ));
-    }
-    let instructions = field_value(&fields, "instructions");
-    if provider == "elevenlabs" && instructions.is_some() {
-        return Ok(bad_request(
-            "ElevenLabs does not accept free-form instructions",
-        ));
-    }
-    let max_chars = field_value(&fields, "max_chars")
-        .and_then(|value| value.parse::<usize>().ok())
-        .unwrap_or(2_000);
-    let provider_max_chars = audio_provider_max_chars(&provider, &model);
-    if max_chars == 0 || max_chars > provider_max_chars {
-        if auto_model {
-            return Ok(bad_request(
-                "ElevenLabs Auto accepts between 1 and 10000 characters per request; pick eleven_flash_v2_5 explicitly for larger values",
-            ));
-        }
-        return Ok(bad_request(&format!(
-            "{provider} accepts between 1 and {provider_max_chars} characters per request"
-        )));
-    }
-    let concurrency = field_value(&fields, "concurrency")
-        .and_then(|value| value.parse::<usize>().ok())
-        .unwrap_or(4)
-        .clamp(1, 16);
-    let gap_chapter_ms = match parse_audio_gap(&fields, "gap_chapter_ms") {
-        Ok(value) => value,
-        Err(error) => return Ok(bad_request(&error)),
-    };
-    let gap_title_ms = match parse_audio_gap(&fields, "gap_title_ms") {
-        Ok(value) => value,
-        Err(error) => return Ok(bad_request(&error)),
-    };
-    let seed = match field_value(&fields, "seed") {
-        Some(value) => match value.parse::<u32>() {
-            Ok(value) => Some(value),
-            Err(_) => return Ok(bad_request("seed must be a valid unsigned 32-bit integer")),
-        },
-        None => None,
-    };
-    let language = match fields.get("language") {
-        Some(value) => {
-            let value = value.trim();
-            if !valid_audio_language(value) {
-                return Ok(bad_request(
-                    "language must be a code such as it, en-US, or pt-BR",
-                ));
-            }
-            Some(value.to_string())
-        }
-        None => None,
-    };
-    let advanced = AudiobookCommandOptions {
-        gap_chapter_ms,
-        gap_title_ms,
-        single: truthy_field(&fields, "single"),
-        loudnorm: truthy_field(&fields, "loudnorm"),
-        seed,
-        language,
-    };
-    let make_m4b = truthy_field(&fields, "m4b");
-    let stitch_audio = truthy_field(&fields, "stitch") || make_m4b;
-    if stitch_audio && format == "pcm" {
-        return Ok(bad_request(
-            "raw PCM cannot be stitched; choose a container format",
-        ));
-    }
-    if make_m4b && !bookforge_audio::ffmpeg_available() {
-        return Ok(bad_request("ffmpeg is required to create an M4B"));
-    }
-    let base_url = field_value(&fields, "base_url");
-    if base_url
-        .as_deref()
-        .is_some_and(|value| !dashboard_audio_base_url_allowed(value))
-    {
-        return Ok(bad_request(
-            "base URL must use HTTPS, except for loopback HTTP endpoints",
-        ));
-    }
-
-    let key_slot = format!("audio:{provider}");
-    let supplied_key = (provider != "mock")
-        .then(|| field_value(&fields, "api_key"))
-        .flatten();
-    let key = if let Some(key) = supplied_key {
-        lock_keys(&state)?.insert(key_slot.clone(), key.clone());
-        Some(key)
-    } else {
-        resolve_audio_provider_key(&state, &provider)?
-    };
-    if provider != "mock"
-        && key.is_none()
-        && !audio_provider_env_has_key(&provider)
-        && !(provider == "openai" && base_url.as_deref().is_some_and(audio_base_url_is_loopback))
-    {
-        return Ok(bad_request("TTS provider API key is required"));
-    }
-
-    let stem = sanitize_component(strip_epub_suffix(&file_name));
-    let sequence = ESTIMATE_SEQ.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-    let id = format!("{}-{sequence}-{stem}", now_ms());
-    let upload_dir = state.upload_dir.clone();
-    std::fs::create_dir_all(&upload_dir)?;
-    let input_path = upload_dir.join(format!("audiobook-{id}.epub"));
-    let out_dir = upload_dir.join(format!("audiobook-{id}"));
-    std::fs::write(&input_path, bytes)?;
-    let inspect_path = input_path.clone();
-    if let Err(error) =
-        tokio::task::spawn_blocking(move || bookforge_epub::inspect_epub(&inspect_path)).await?
-    {
-        let _ = std::fs::remove_file(&input_path);
-        return Ok(bad_request(&format!("could not read EPUB: {error}")));
-    }
-    std::fs::create_dir_all(&out_dir)?;
-    write_audio_process_state(&out_dir, "starting", None, None, auto_model)?;
-
-    let exe = std::env::current_exe()?;
-    let mut command = tokio::process::Command::new(exe);
-    let api_key_env = (provider != "mock" && key.is_some())
-        .then(|| audio_provider_key_env(&provider).expect("audio provider was validated"));
-    command.args(audiobook_command_args(
-        &input_path,
-        &out_dir,
-        &provider,
-        (!auto_model).then_some(model.as_str()),
-        &voice,
-        &format,
-        speed,
-        max_chars,
-        concurrency,
-        instructions.as_deref(),
-        base_url.as_deref(),
-        make_m4b,
-        stitch_audio,
-        api_key_env,
-        &advanced,
-    ));
-    configure_dashboard_child_environment(&mut command, api_key_env.zip(key.as_deref()));
-
-    let mut child = command
-        .spawn()
-        .context("failed to spawn audiobook process")?;
-    let pid = child.id();
-    write_audio_process_state(&out_dir, "running", pid, None, auto_model)?;
-    let cancel = tokio_util::sync::CancellationToken::new();
-    state
-        .audio_cancels
-        .lock()
-        .map_err(|_| anyhow::anyhow!("audiobook cancellation registry is unavailable"))?
-        .insert(id.clone(), cancel.clone());
-    let monitor_dir = out_dir.clone();
-    let monitor_id = id.clone();
-    let cancel_registry = state.audio_cancels.clone();
-    tokio::spawn(async move {
-        let operation_state = tokio::select! {
-            status = child.wait() => match status {
-                Ok(status) if status.success() => ("succeeded", None),
-                Ok(status) => ("failed", Some(format!("audiobook process exited with {status}"))),
-                Err(error) => ("failed", Some(format!("could not wait for audiobook process: {error}"))),
-            },
-            _ = cancel.cancelled() => {
-                let _ = child.kill().await;
-                let _ = child.wait().await;
-                ("cancelled", None)
-            }
-        };
-        let _ = write_audio_process_state(
-            &monitor_dir,
-            operation_state.0,
-            pid,
-            operation_state.1.as_deref(),
-            auto_model,
-        );
-        if let Ok(mut registry) = cancel_registry.lock() {
-            registry.remove(&monitor_id);
-        }
-    });
-
-    Ok(Json(json!({
-        "ok": true,
-        "id": id,
-        "input_path": input_path.display().to_string(),
-        "out_dir": out_dir.display().to_string(),
-        "provider": provider,
-        "model": model,
-        "voice": voice,
-        "pid": pid,
-    }))
-    .into_response())
-}
-
-#[derive(Debug, Default)]
-struct AudiobookCommandOptions {
-    gap_chapter_ms: Option<u32>,
-    gap_title_ms: Option<u32>,
-    single: bool,
-    loudnorm: bool,
-    seed: Option<u32>,
-    language: Option<String>,
-}
-
-fn clamp_audio_gap(value: u32) -> u32 {
-    value.min(10_000)
-}
-
-fn parse_audio_gap(
-    fields: &HashMap<String, String>,
-    name: &str,
-) -> std::result::Result<Option<u32>, String> {
-    let Some(value) = fields.get(name) else {
-        return Ok(None);
-    };
-    let value = value
-        .trim()
-        .parse::<u32>()
-        .map_err(|_| format!("{name} must be a non-negative integer"))?;
-    Ok(Some(clamp_audio_gap(value)))
-}
-
-fn valid_audio_language(value: &str) -> bool {
-    let mut parts = value.split('-');
-    let Some(primary) = parts.next() else {
-        return false;
-    };
-    if !(2..=3).contains(&primary.len()) || !primary.bytes().all(|byte| byte.is_ascii_alphabetic())
-    {
-        return false;
-    }
-    parts.all(|part| {
-        (2..=8).contains(&part.len()) && part.bytes().all(|byte| byte.is_ascii_alphanumeric())
-    })
-}
-
-#[allow(clippy::too_many_arguments)]
-fn audiobook_command_args(
-    input_path: &Path,
-    out_dir: &Path,
-    provider: &str,
-    model: Option<&str>,
-    voice: &str,
-    format: &str,
-    speed: f32,
-    max_chars: usize,
-    concurrency: usize,
-    instructions: Option<&str>,
-    base_url: Option<&str>,
-    make_m4b: bool,
-    stitch_audio: bool,
-    api_key_env: Option<&str>,
-    advanced: &AudiobookCommandOptions,
-) -> Vec<OsString> {
-    let mut args = vec![
-        "audiobook".into(),
-        input_path.as_os_str().to_owned(),
-        "--out".into(),
-        out_dir.as_os_str().to_owned(),
-        "--provider".into(),
-        provider.into(),
-    ];
-    if let Some(model) = model.filter(|value| !value.is_empty()) {
-        args.extend([OsString::from("--model"), model.into()]);
-    }
-    args.extend([
-        "--voice".into(),
-        voice.into(),
-        "--format".into(),
-        format.into(),
-        "--speed".into(),
-        speed.to_string().into(),
-        "--max-chars".into(),
-        max_chars.to_string().into(),
-        "--concurrency".into(),
-        concurrency.to_string().into(),
-        "--ui".into(),
-        "quiet".into(),
-    ]);
-    if let Some(instructions) = instructions {
-        args.extend([OsString::from("--instructions"), instructions.into()]);
-    }
-    if let Some(base_url) = base_url {
-        args.extend([OsString::from("--base-url"), base_url.into()]);
-    }
-    if make_m4b {
-        args.push("--m4b".into());
-    } else if stitch_audio {
-        args.push("--stitch".into());
-    }
-    if let Some(env) = api_key_env {
-        args.extend([OsString::from("--api-key-env"), env.into()]);
-    }
-    if let Some(gap) = advanced.gap_chapter_ms {
-        args.extend([OsString::from("--gap-chapter-ms"), gap.to_string().into()]);
-    }
-    if let Some(gap) = advanced.gap_title_ms {
-        args.extend([OsString::from("--gap-title-ms"), gap.to_string().into()]);
-    }
-    if advanced.single {
-        args.push("--single".into());
-    }
-    if advanced.loudnorm {
-        args.push("--loudnorm".into());
-    }
-    if let Some(seed) = advanced.seed {
-        args.extend([OsString::from("--seed"), seed.to_string().into()]);
-    }
-    if let Some(language) = advanced.language.as_deref() {
-        args.extend([OsString::from("--language"), language.into()]);
-    }
-    args
-}
-
-fn resolved_model_from_synthesis_id(synthesis_id: &str) -> Option<&str> {
-    let (_, model) = synthesis_id.rsplit_once(':')?;
-    (!model.is_empty()).then_some(model)
-}
-
-async fn list_audiobooks(
-    State(state): State<AppState>,
-) -> Result<Json<Vec<serde_json::Value>>, AppError> {
-    let upload_dir = state.upload_dir.clone();
-    let items =
-        tokio::task::spawn_blocking(move || list_audiobook_summaries(&upload_dir)).await??;
-    Ok(Json(items))
-}
-
-fn list_audiobook_summaries(upload_dir: &Path) -> Result<Vec<serde_json::Value>> {
-    let entries = match std::fs::read_dir(upload_dir) {
-        Ok(entries) => entries,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(Vec::new()),
-        Err(error) => return Err(error.into()),
-    };
-    let mut items = Vec::new();
-    for entry in entries {
-        let entry = entry?;
-        if !entry.file_type()?.is_dir() {
-            continue;
-        }
-        let name = entry.file_name();
-        let Some(id) = name
-            .to_str()
-            .and_then(|name| name.strip_prefix("audiobook-"))
-            .filter(|id| valid_audiobook_id(id))
-        else {
-            continue;
-        };
-        let out_dir = entry.path();
-        if !out_dir.join("manifest.json").is_file() && !out_dir.join("process.json").is_file() {
-            continue;
-        }
-        if let Some(payload) = read_audiobook_payload(upload_dir, id) {
-            let process_updated = payload
-                .get("process")
-                .and_then(|process| process.get("updated_at_ms"))
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(0);
-            let manifest_updated = payload
-                .get("updated_at_ms")
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(0);
-            items.push(json!({
-                "id": payload.get("id"),
-                "title": payload.get("title"),
-                "status": payload.get("status"),
-                "process_status": payload
-                    .get("process")
-                    .and_then(|process| process.get("status")),
-                "input_path": payload.get("input_path"),
-                "out_dir": payload.get("out_dir"),
-                "synthesis_id": payload.get("synthesis_id"),
-                "voice": payload.get("voice"),
-                "chapters": payload.get("chapters"),
-                "completed_chunks": payload.get("completed_chunks"),
-                "total_chunks": payload
-                    .get("chunks")
-                    .and_then(serde_json::Value::as_array)
-                    .map(Vec::len)
-                    .unwrap_or(0),
-                "artifact": payload.get("artifact"),
-                "warnings": payload.get("warnings"),
-                "updated_at_ms": manifest_updated.max(process_updated),
-            }));
-        }
-    }
-    items.sort_by_key(|item| {
-        std::cmp::Reverse(
-            item.get("updated_at_ms")
-                .and_then(serde_json::Value::as_u64)
-                .unwrap_or(0),
-        )
-    });
-    Ok(items)
-}
-
-fn audiobook_warnings(process: &serde_json::Value) -> Vec<serde_json::Value> {
-    match process.get("warnings") {
-        Some(serde_json::Value::Array(warnings)) => warnings.clone(),
-        Some(serde_json::Value::String(warning)) if !warning.is_empty() => {
-            vec![json!(warning)]
-        }
-        Some(serde_json::Value::Object(warning)) if !warning.is_empty() => {
-            vec![serde_json::Value::Object(warning.clone())]
-        }
-        _ => Vec::new(),
-    }
-}
-
-fn read_audiobook_payload(upload_dir: &Path, id: &str) -> Option<serde_json::Value> {
-    let out_dir = upload_dir.join(format!("audiobook-{id}"));
-    if !out_dir.is_dir() {
-        return None;
-    }
-    let process: serde_json::Value = std::fs::read(out_dir.join("process.json"))
-        .ok()
-        .and_then(|bytes| serde_json::from_slice(&bytes).ok())
-        .unwrap_or_else(|| json!({"status": "starting"}));
-    let mut payload: serde_json::Value = std::fs::read(out_dir.join("manifest.json"))
-        .ok()
-        .and_then(|bytes| serde_json::from_slice(&bytes).ok())
-        .unwrap_or_else(|| {
-            json!({
-                "status": "starting",
-                "completed_chunks": 0,
-                "chunks": [],
-            })
-        });
-    let resolved_model = payload
-        .get("synthesis_id")
-        .and_then(serde_json::Value::as_str)
-        .and_then(resolved_model_from_synthesis_id)
-        .map(str::to_string);
-    let warnings = audiobook_warnings(&process);
-    let object = payload.as_object_mut()?;
-    object.insert("id".to_string(), json!(id));
-    object.insert(
-        "input_path".to_string(),
-        json!(
-            upload_dir
-                .join(format!("audiobook-{id}.epub"))
-                .display()
-                .to_string()
-        ),
-    );
-    object.insert("out_dir".to_string(), json!(out_dir.display().to_string()));
-    object.insert("process".to_string(), process.clone());
-    object.insert("warnings".to_string(), json!(warnings));
-    object.insert("resolved_model".to_string(), json!(resolved_model));
-    match process.get("status").and_then(serde_json::Value::as_str) {
-        Some("succeeded") => {
-            object.insert("status".to_string(), json!("succeeded"));
-        }
-        Some("failed") => {
-            object.insert("status".to_string(), json!("failed"));
-            object.insert(
-                "error".to_string(),
-                process
-                    .get("error")
-                    .cloned()
-                    .unwrap_or_else(|| json!("audiobook process failed")),
-            );
-        }
-        Some("cancelled") => {
-            object.insert("status".to_string(), json!("cancelled"));
-        }
-        _ => {}
-    }
-    if out_dir.join("audiobook.m4b").is_file() {
-        object.insert(
-            "artifact".to_string(),
-            json!(out_dir.join("audiobook.m4b").display().to_string()),
-        );
-    }
-    Some(payload)
-}
-
-async fn audiobook_status(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-) -> Result<Response, AppError> {
-    if !valid_audiobook_id(&id) {
-        return Ok(bad_request("invalid audiobook operation id"));
-    }
-    match read_audiobook_payload(&state.upload_dir, &id) {
-        Some(payload) => Ok(Json(payload).into_response()),
-        None => Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({"error": "no such audiobook operation"})),
-        )
-            .into_response()),
-    }
-}
-
-async fn cancel_audiobook(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    AxumPath(id): AxumPath<String>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    if !valid_audiobook_id(&id) {
-        return Ok(bad_request("invalid audiobook operation id"));
-    }
-    let cancel = state
-        .audio_cancels
-        .lock()
-        .map_err(|_| anyhow::anyhow!("audiobook cancellation registry is unavailable"))?
-        .get(&id)
-        .cloned();
-    if let Some(cancel) = cancel {
-        cancel.cancel();
-        return Ok(Json(json!({"ok": true, "status": "cancelling"})).into_response());
-    }
-
-    let out_dir = state.upload_dir.join(format!("audiobook-{id}"));
-    let process: serde_json::Value = match std::fs::read(out_dir.join("process.json"))
-        .ok()
-        .and_then(|bytes| serde_json::from_slice(&bytes).ok())
-    {
-        Some(process) => process,
-        None => {
-            return Ok((
-                StatusCode::CONFLICT,
-                Json(json!({"error": "audiobook operation has no running process"})),
-            )
-                .into_response());
-        }
-    };
-    if !matches!(
-        process.get("status").and_then(serde_json::Value::as_str),
-        Some("starting" | "running")
-    ) {
-        return Ok((
-            StatusCode::CONFLICT,
-            Json(json!({"error": "audiobook operation is not running"})),
-        )
-            .into_response());
-    }
-    let Some(pid) = process
-        .get("pid")
-        .and_then(serde_json::Value::as_u64)
-        .and_then(|pid| u32::try_from(pid).ok())
-        .filter(|pid| *pid != 0)
-    else {
-        return Ok((
-            StatusCode::CONFLICT,
-            Json(json!({"error": "audiobook operation has no running process"})),
-        )
-            .into_response());
-    };
-    if let Err(error) = terminate_restarted_audiobook(&state, pid).await {
-        return Ok((
-            StatusCode::CONFLICT,
-            Json(json!({"error": format!("could not cancel audiobook process: {error}")})),
-        )
-            .into_response());
-    }
-    let auto_model = process
-        .get("auto_model")
-        .and_then(serde_json::Value::as_bool)
-        .unwrap_or(false);
-    write_audio_process_state(&out_dir, "cancelled", Some(pid), None, auto_model)?;
-    Ok(Json(json!({"ok": true, "status": "cancelled"})).into_response())
-}
-
-async fn terminate_restarted_audiobook(state: &AppState, pid: u32) -> Result<()> {
-    #[cfg(test)]
-    if let Some(cancelled) = &state.audio_restart_cancels {
-        cancelled
-            .lock()
-            .map_err(|_| anyhow::anyhow!("test cancellation recorder is unavailable"))?
-            .push(pid);
-        return Ok(());
-    }
-    #[cfg(not(test))]
-    let _ = state;
-
-    #[cfg(windows)]
-    let status = tokio::process::Command::new("taskkill")
-        .args(["/PID", &pid.to_string(), "/T", "/F"])
-        .status()
-        .await
-        .context("failed to run taskkill")?;
-    #[cfg(unix)]
-    let status = tokio::process::Command::new("kill")
-        .args(["-TERM", &pid.to_string()])
-        .status()
-        .await
-        .context("failed to run kill")?;
-    #[cfg(not(any(windows, unix)))]
-    anyhow::bail!("process cancellation is unsupported on this platform");
-
-    if !status.success() {
-        anyhow::bail!("process signalling exited with {status}");
-    }
-    Ok(())
-}
-
-#[derive(Deserialize)]
-struct ArtifactQuery {
-    disposition: Option<String>,
-}
-
-#[derive(Debug, PartialEq, Eq)]
-enum ArtifactByteRange {
-    Full,
-    Partial { start: u64, end: u64 },
-    Unsatisfiable,
-}
-
-fn parse_artifact_range(headers: &HeaderMap, length: u64) -> ArtifactByteRange {
-    let Some(value) = headers.get(axum::http::header::RANGE) else {
-        return ArtifactByteRange::Full;
-    };
-    let Ok(value) = value.to_str() else {
-        return ArtifactByteRange::Unsatisfiable;
-    };
-    let Some(spec) = value.trim().strip_prefix("bytes=") else {
-        return ArtifactByteRange::Unsatisfiable;
-    };
-    if spec.contains(',') {
-        return ArtifactByteRange::Unsatisfiable;
-    }
-    let Some((start, end)) = spec.split_once('-') else {
-        return ArtifactByteRange::Unsatisfiable;
-    };
-    if length == 0 {
-        return ArtifactByteRange::Unsatisfiable;
-    }
-    if start.is_empty() {
-        let Ok(suffix_length) = end.parse::<u64>() else {
-            return ArtifactByteRange::Unsatisfiable;
-        };
-        if suffix_length == 0 {
-            return ArtifactByteRange::Unsatisfiable;
-        }
-        let start = length.saturating_sub(suffix_length.min(length));
-        return ArtifactByteRange::Partial {
-            start,
-            end: length - 1,
-        };
-    }
-    let Ok(start) = start.parse::<u64>() else {
-        return ArtifactByteRange::Unsatisfiable;
-    };
-    if start >= length {
-        return ArtifactByteRange::Unsatisfiable;
-    }
-    let end = if end.is_empty() {
-        length - 1
-    } else {
-        let Ok(end) = end.parse::<u64>() else {
-            return ArtifactByteRange::Unsatisfiable;
-        };
-        end.min(length - 1)
-    };
-    if end < start {
-        return ArtifactByteRange::Unsatisfiable;
-    }
-    ArtifactByteRange::Partial { start, end }
-}
-
-async fn audiobook_artifact(
-    State(state): State<AppState>,
-    AxumPath(id): AxumPath<String>,
-    Query(query): Query<ArtifactQuery>,
-    headers: HeaderMap,
-) -> Result<Response, AppError> {
-    if !valid_audiobook_id(&id) {
-        return Ok(bad_request("invalid audiobook operation id"));
-    }
-    let out_dir = state.upload_dir.join(format!("audiobook-{id}"));
-    let m4b_path = out_dir.join("audiobook.m4b");
-    let (path, content_type, download_name) = if m4b_path.is_file() {
-        (m4b_path, "audio/mp4", "audiobook.m4b")
-    } else {
-        let archive_dir = out_dir.clone();
-        let archive =
-            tokio::task::spawn_blocking(move || ensure_audio_download_zip(&archive_dir)).await??;
-        (archive, "application/zip", "audiobook-audio.zip")
-    };
-    let Ok(mut file) = tokio::fs::File::open(&path).await else {
-        return Ok((
-            StatusCode::NOT_FOUND,
-            Json(json!({"error": "audiobook artifact is not available"})),
-        )
-            .into_response());
-    };
-    let inline = query.disposition.as_deref() == Some("inline");
-    let content_disposition = match (inline, download_name.ends_with(".m4b")) {
-        (true, true) => "inline; filename=\"audiobook.m4b\"",
-        (true, false) => "inline; filename=\"audiobook-audio.zip\"",
-        (false, true) => "attachment; filename=\"audiobook.m4b\"",
-        (false, false) => "attachment; filename=\"audiobook-audio.zip\"",
-    };
-    let length = file.metadata().await?.len();
-    let requested = parse_artifact_range(&headers, length);
-    if requested == ArtifactByteRange::Unsatisfiable {
-        return Ok(Response::builder()
-            .status(StatusCode::RANGE_NOT_SATISFIABLE)
-            .header(axum::http::header::ACCEPT_RANGES, "bytes")
-            .header(
-                axum::http::header::CONTENT_RANGE,
-                format!("bytes */{length}"),
-            )
-            .body(axum::body::Body::empty())
-            .context("failed to build range error response")?);
-    }
-
-    let mut builder = Response::builder()
-        .header(axum::http::header::CONTENT_TYPE, content_type)
-        .header(axum::http::header::CONTENT_DISPOSITION, content_disposition)
-        .header(axum::http::header::ACCEPT_RANGES, "bytes");
-    match requested {
-        ArtifactByteRange::Full => Ok(builder
-            .header(axum::http::header::CONTENT_LENGTH, length)
-            .body(axum::body::Body::from_stream(
-                tokio_util::io::ReaderStream::new(file),
-            ))
-            .context("failed to build artifact response")?),
-        ArtifactByteRange::Partial { start, end } => {
-            use tokio::io::{AsyncReadExt, AsyncSeekExt};
-
-            file.seek(std::io::SeekFrom::Start(start)).await?;
-            let partial_length = end - start + 1;
-            builder = builder
-                .status(StatusCode::PARTIAL_CONTENT)
-                .header(axum::http::header::CONTENT_LENGTH, partial_length)
-                .header(
-                    axum::http::header::CONTENT_RANGE,
-                    format!("bytes {start}-{end}/{length}"),
-                );
-            Ok(builder
-                .body(axum::body::Body::from_stream(
-                    tokio_util::io::ReaderStream::new(file.take(partial_length)),
-                ))
-                .context("failed to build partial artifact response")?)
-        }
-        ArtifactByteRange::Unsatisfiable => unreachable!("handled above"),
-    }
-}
-
-fn ensure_audio_download_zip(out_dir: &std::path::Path) -> Result<PathBuf> {
-    use std::io::{Read, Write};
-
-    let archive_path = out_dir.join("audiobook-audio.zip");
-    if archive_path.is_file() {
-        return Ok(archive_path);
-    }
-    let manifest_path = out_dir.join("manifest.json");
-    let manifest: bookforge_audio::AudiobookManifest = serde_json::from_slice(
-        &std::fs::read(&manifest_path)
-            .with_context(|| format!("failed to read {}", manifest_path.display()))?,
-    )?;
-    let staged = out_dir.join("audiobook-audio.zip.tmp");
-    let _ = std::fs::remove_file(&staged);
-    let file = std::fs::File::create(&staged)?;
-    let mut archive = zip::ZipWriter::new(file);
-    let options =
-        zip::write::SimpleFileOptions::default().compression_method(zip::CompressionMethod::Stored);
-    archive.start_file("manifest.json", options)?;
-    archive.write_all(&std::fs::read(&manifest_path)?)?;
-    for chunk in &manifest.chunks {
-        let name = std::path::Path::new(&chunk.file);
-        if name.file_name().and_then(|value| value.to_str()) != Some(chunk.file.as_str()) {
-            anyhow::bail!("unsafe audio filename in manifest");
-        }
-        let source = out_dir.join(&chunk.file);
-        let mut source_file = std::fs::File::open(&source)
-            .with_context(|| format!("failed to read audio part {}", source.display()))?;
-        archive.start_file(&chunk.file, options)?;
-        let mut buffer = [0u8; 64 * 1024];
-        loop {
-            let read = source_file.read(&mut buffer)?;
-            if read == 0 {
-                break;
-            }
-            archive.write_all(&buffer[..read])?;
-        }
-    }
-    archive.finish()?;
-    if let Err(error) = std::fs::rename(&staged, &archive_path) {
-        let _ = std::fs::remove_file(&staged);
-        if !archive_path.is_file() {
-            return Err(error.into());
-        }
-    }
-    Ok(archive_path)
-}
-
-fn valid_audiobook_id(id: &str) -> bool {
-    !id.is_empty()
-        && id.len() <= 160
-        && id
-            .bytes()
-            .all(|byte| byte.is_ascii_alphanumeric() || matches!(byte, b'-' | b'_'))
-}
-
-fn write_audio_process_state(
-    out_dir: &std::path::Path,
-    status: &str,
-    pid: Option<u32>,
-    error: Option<&str>,
-    auto_model: bool,
-) -> Result<()> {
-    let path = out_dir.join("process.json");
-    let temp = out_dir.join("process.part.tmp");
-    let warnings = std::fs::read(&path)
-        .ok()
-        .and_then(|bytes| serde_json::from_slice::<serde_json::Value>(&bytes).ok())
-        .and_then(|process| process.get("warnings").cloned());
-    let mut process = json!({
-        "status": status,
-        "pid": pid,
-        "error": error,
-        "auto_model": auto_model,
-        "updated_at_ms": now_ms(),
-    });
-    if let (Some(object), Some(warnings)) = (process.as_object_mut(), warnings) {
-        object.insert("warnings".to_string(), warnings);
-    }
-    let bytes = serde_json::to_vec_pretty(&process)?;
-    std::fs::write(&temp, bytes)?;
-    if path.exists() {
-        let _ = std::fs::remove_file(&path);
-    }
-    std::fs::rename(temp, path)?;
-    Ok(())
-}
 
 async fn child_exit_status_after(
     child: &mut tokio::process::Child,
@@ -2370,488 +405,6 @@ async fn child_exit_status_after(
         .context("failed to check translation process status")
 }
 
-/// Estimate tokens and cost for an uploaded EPUB before the user commits to a
-/// run. Shares [`estimate_epub`](super::estimate::estimate_epub) with the CLI
-/// `estimate` command. The upload is written to a temp file (EPUB parsing reads
-/// from disk) and removed immediately after.
-async fn estimate_translate(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    mut multipart: Multipart,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-
-    let mut file_bytes: Option<Vec<u8>> = None;
-    let mut fields: HashMap<String, String> = HashMap::new();
-    while let Some(field) = multipart.next_field().await? {
-        let name = field.name().unwrap_or_default().to_string();
-        if name == "file" {
-            file_bytes = Some(field.bytes().await?.to_vec());
-        } else {
-            fields.insert(name, field.text().await?);
-        }
-    }
-
-    let Some(bytes) = file_bytes.filter(|b| !b.is_empty()) else {
-        return Ok(bad_request("upload an EPUB file"));
-    };
-    let provider = field_value(&fields, "provider").unwrap_or_else(|| "mock".to_string());
-    if !supported_provider(&provider) {
-        return Ok(bad_request("unsupported provider"));
-    }
-    let model = field_value(&fields, "model");
-    let target = field_value(&fields, "target").unwrap_or_else(|| "Italian".to_string());
-
-    let result = tokio::task::spawn_blocking(move || {
-        let seq = ESTIMATE_SEQ.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
-        let path = std::env::temp_dir().join(format!(
-            "bookforge-estimate-{}-{}-{}.epub",
-            std::process::id(),
-            now_ms(),
-            seq
-        ));
-        std::fs::write(&path, &bytes)?;
-        let outcome =
-            super::estimate::estimate_epub(&path, &target, &provider, model.as_deref(), None);
-        let _ = std::fs::remove_file(&path);
-        outcome
-    })
-    .await?;
-
-    match result {
-        Ok(est) => Ok(Json(json!({
-            "segments": est.segments,
-            "input_tokens": est.input_tokens,
-            "output_tokens": est.output_tokens,
-            "model": est.model,
-            "cost_usd": est.cost_usd,
-        }))
-        .into_response()),
-        Err(err) => Ok(bad_request(&format!("could not estimate: {err}"))),
-    }
-}
-
-/// Report which providers already have a usable key — either remembered in this
-/// session or present in the server's environment — so the UI only prompts when
-/// a key is actually needed. Never returns key material.
-async fn provider_status(
-    State(state): State<AppState>,
-) -> Result<Json<serde_json::Value>, AppError> {
-    let remembered = lock_keys(&state)?;
-    let mut status = serde_json::Map::new();
-    for (provider, env) in PROVIDER_KEY_ENVS {
-        let configured = remembered.contains_key(*provider)
-            || std::env::var(env).map(|v| !v.is_empty()).unwrap_or(false);
-        status.insert((*provider).to_string(), json!(configured));
-    }
-    for (provider, env) in AUDIO_PROVIDER_KEY_ENVS {
-        let configured = remembered.contains_key(&format!("audio:{provider}"))
-            || std::env::var(env)
-                .map(|value| !value.is_empty())
-                .unwrap_or(false);
-        status.insert(format!("audio:{provider}"), json!(configured));
-    }
-    Ok(Json(serde_json::Value::Object(status)))
-}
-
-#[derive(Deserialize)]
-struct AudioVoicesQuery {
-    provider: Option<String>,
-}
-
-async fn audio_voices(
-    State(state): State<AppState>,
-    Query(query): Query<AudioVoicesQuery>,
-) -> Result<Response, AppError> {
-    if query.provider.as_deref() != Some("elevenlabs") {
-        return Ok(bad_request(
-            "voice listing is available for ElevenLabs only",
-        ));
-    }
-
-    let Some(api_key) = resolve_audio_provider_key(&state, "elevenlabs")? else {
-        return Ok((
-            StatusCode::CONFLICT,
-            Json(json!({ "error": "ElevenLabs API key is not configured" })),
-        )
-            .into_response());
-    };
-
-    if let Some(voices) = state
-        .elevenlabs_voices
-        .lock()
-        .map_err(|_| anyhow::anyhow!("ElevenLabs voice cache is unavailable"))?
-        .as_ref()
-        .filter(|cache| cache.fetched_at.elapsed() < ELEVENLABS_VOICE_CACHE_TTL)
-        .map(|cache| cache.voices.clone())
-    {
-        return Ok(Json(json!({ "voices": voices })).into_response());
-    }
-
-    let voices = match bookforge_audio::list_elevenlabs_voices(
-        ELEVENLABS_BASE_URL,
-        &api_key,
-        ELEVENLABS_VOICE_TIMEOUT_SECONDS,
-    )
-    .await
-    {
-        Ok(voices) => voices,
-        Err(_) => {
-            return Ok((
-                StatusCode::BAD_GATEWAY,
-                Json(json!({ "error": "could not load ElevenLabs voices" })),
-            )
-                .into_response());
-        }
-    };
-    *state
-        .elevenlabs_voices
-        .lock()
-        .map_err(|_| anyhow::anyhow!("ElevenLabs voice cache is unavailable"))? =
-        Some(ElevenLabsVoiceCache {
-            fetched_at: Instant::now(),
-            voices: voices.clone(),
-        });
-
-    Ok(Json(json!({ "voices": voices })).into_response())
-}
-
-// ---------------------------------------------------------------------------
-// Glossary (wraps the `glossary` command's JobStore methods)
-// ---------------------------------------------------------------------------
-
-fn parse_glossary_scope(value: &str) -> GlossaryScopeKind {
-    match value {
-        "series" => GlossaryScopeKind::Series,
-        "book" => GlossaryScopeKind::Book,
-        _ => GlossaryScopeKind::Global,
-    }
-}
-
-fn parse_glossary_category(value: &str) -> GlossaryCategory {
-    match value {
-        "person" => GlossaryCategory::Person,
-        "place" => GlossaryCategory::Place,
-        "object" => GlossaryCategory::Object,
-        "invented" => GlossaryCategory::Invented,
-        "style" => GlossaryCategory::Style,
-        "phrase" => GlossaryCategory::Phrase,
-        _ => GlossaryCategory::Other,
-    }
-}
-
-fn glossary_term_json(term: &GlossaryTerm) -> serde_json::Value {
-    json!({
-        "id": term.id,
-        "source": term.source_text,
-        "target": term.target_text,
-        "category": term.category.as_str(),
-        "scope": term.scope_kind.as_str(),
-        "scope_id": term.scope_id,
-        "source_language": term.source_language,
-        "target_language": term.target_language,
-        "always_active": term.always_active,
-        "case_sensitive": term.case_sensitive,
-        "notes": term.notes,
-    })
-}
-
-/// List glossary terms, optionally filtered by `source`/`target` language,
-/// `scope` (global/series/book) and `scope_id`.
-async fn list_glossary(
-    Query(params): Query<HashMap<String, String>>,
-    State(state): State<AppState>,
-) -> Result<Json<Vec<serde_json::Value>>, AppError> {
-    let store_path = state.store_path.clone();
-    let items = tokio::task::spawn_blocking(move || -> Result<Vec<serde_json::Value>> {
-        let store = JobStore::open(store_path)?;
-        let terms = store.list_glossary_terms(GlossaryFilter {
-            scope_kind: params.get("scope").map(|value| parse_glossary_scope(value)),
-            scope_id: params
-                .get("scope_id")
-                .map(String::as_str)
-                .filter(|value| !value.is_empty()),
-            source_language: params
-                .get("source")
-                .map(String::as_str)
-                .filter(|value| !value.is_empty()),
-            target_language: params
-                .get("target")
-                .map(String::as_str)
-                .filter(|value| !value.is_empty()),
-            active_only: false,
-        })?;
-        Ok(terms.iter().map(glossary_term_json).collect())
-    })
-    .await??;
-    Ok(Json(items))
-}
-
-#[derive(Deserialize)]
-struct GlossaryAddRequest {
-    source: String,
-    target: String,
-    source_language: String,
-    target_language: String,
-    #[serde(default)]
-    category: Option<String>,
-    #[serde(default)]
-    scope: Option<String>,
-    #[serde(default)]
-    scope_id: Option<String>,
-    #[serde(default)]
-    notes: Option<String>,
-    #[serde(default)]
-    case_sensitive: bool,
-    #[serde(default)]
-    always_active: bool,
-}
-
-async fn add_glossary(
-    State(state): State<AppState>,
-    headers: HeaderMap,
-    Json(req): Json<GlossaryAddRequest>,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    if req.source.trim().is_empty() || req.target.trim().is_empty() {
-        return Ok(bad_request("source term and translation are required"));
-    }
-    if req.source_language.trim().is_empty() || req.target_language.trim().is_empty() {
-        return Ok(bad_request("source and target languages are required"));
-    }
-    let scope_kind = req
-        .scope
-        .as_deref()
-        .map(parse_glossary_scope)
-        .unwrap_or(GlossaryScopeKind::Global);
-    let scope_id = if scope_kind == GlossaryScopeKind::Global {
-        None
-    } else {
-        req.scope_id.filter(|value| !value.trim().is_empty())
-    };
-    if scope_kind != GlossaryScopeKind::Global && scope_id.is_none() {
-        return Ok(bad_request("scope_id is required for series/book scope"));
-    }
-
-    let term = GlossaryTerm {
-        id: None,
-        scope_kind,
-        scope_id,
-        source_text: req.source.trim().to_string(),
-        target_text: req.target.trim().to_string(),
-        category: req
-            .category
-            .as_deref()
-            .map(parse_glossary_category)
-            .unwrap_or(GlossaryCategory::Other),
-        notes: req.notes.filter(|value| !value.trim().is_empty()),
-        case_sensitive: req.case_sensitive,
-        always_active: req.always_active,
-        status: GlossaryStatus::UserSeeded,
-        source_language: req.source_language.trim().to_string(),
-        target_language: req.target_language.trim().to_string(),
-        source_count: 0,
-    };
-
-    let store_path = state.store_path.clone();
-    let id = tokio::task::spawn_blocking(move || -> Result<i64> {
-        let store = JobStore::open(store_path)?;
-        Ok(store.add_glossary_term(&term)?)
-    })
-    .await??;
-    Ok(Json(json!({ "id": id })).into_response())
-}
-
-async fn remove_glossary(
-    AxumPath(id): AxumPath<i64>,
-    State(state): State<AppState>,
-    headers: HeaderMap,
-) -> Result<Response, AppError> {
-    if let Some(response) = reject_mutation(&headers, &state) {
-        return Ok(response);
-    }
-    let store_path = state.store_path.clone();
-    let removed = tokio::task::spawn_blocking(move || -> Result<usize> {
-        let store = JobStore::open(store_path)?;
-        Ok(store.remove_glossary_term(id)?)
-    })
-    .await??;
-    Ok(Json(json!({ "removed": removed })).into_response())
-}
-
-fn supported_provider(provider: &str) -> bool {
-    provider == "mock" || provider_key_env(provider).is_some()
-}
-
-fn audio_provider_key_env(provider: &str) -> Option<&'static str> {
-    AUDIO_PROVIDER_KEY_ENVS
-        .iter()
-        .find_map(|(known, env)| (*known == provider).then_some(*env))
-}
-
-fn audio_provider_env_has_key(provider: &str) -> bool {
-    audio_provider_key_env(provider)
-        .and_then(|env| std::env::var(env).ok())
-        .is_some_and(|value| !value.is_empty())
-}
-
-fn resolve_audio_provider_key(state: &AppState, provider: &str) -> Result<Option<String>> {
-    let key_slot = format!("audio:{provider}");
-    if let Some(key) = lock_keys(state)?.get(&key_slot).cloned() {
-        return Ok(Some(key));
-    }
-    Ok(audio_provider_key_env(provider)
-        .and_then(|env| std::env::var(env).ok())
-        .filter(|value| !value.is_empty()))
-}
-
-fn audio_base_url_is_loopback(value: &str) -> bool {
-    reqwest::Url::parse(value)
-        .ok()
-        .and_then(|url| url.host_str().map(str::to_string))
-        .is_some_and(|host| matches!(host.as_str(), "localhost" | "127.0.0.1" | "::1"))
-}
-
-fn dashboard_audio_base_url_allowed(value: &str) -> bool {
-    reqwest::Url::parse(value).is_ok_and(|url| {
-        url.host().is_some()
-            && url.username().is_empty()
-            && url.password().is_none()
-            && url.query().is_none()
-            && url.fragment().is_none()
-            && (url.scheme() == "https"
-                || (url.scheme() == "http" && audio_base_url_is_loopback(value)))
-    })
-}
-
-fn dashboard_base_url_uses_https(base_url: &str) -> bool {
-    reqwest::Url::parse(base_url).is_ok_and(|url| url.scheme() == "https" && url.host().is_some())
-}
-
-fn provider_key_env(provider: &str) -> Option<&'static str> {
-    PROVIDER_KEY_ENVS
-        .iter()
-        .find_map(|(known, env)| (*known == provider).then_some(*env))
-}
-
-/// Resolve a translation provider key without persisting it.
-///
-/// A key supplied by the browser replaces the session's remembered key for that
-/// provider. Otherwise the session key wins over the one expected in the serve
-/// process's environment. Callers still choose the env-var name injected into
-/// the scrubbed child, which lets resume honor the run snapshot.
-fn resolve_dashboard_provider_key(
-    state: &AppState,
-    provider: &str,
-    supplied_key: Option<String>,
-    expected_env: Option<&str>,
-) -> Result<Option<String>> {
-    let supplied_key = supplied_key.and_then(|key| {
-        let key = key.trim();
-        (!key.is_empty()).then(|| key.to_string())
-    });
-    if let Some(supplied_key) = supplied_key {
-        lock_keys(state)?.insert(provider.to_string(), supplied_key.clone());
-        return Ok(Some(supplied_key));
-    }
-    if let Some(remembered_key) = lock_keys(state)?.get(provider).cloned() {
-        return Ok(Some(remembered_key));
-    }
-    Ok(expected_env
-        .and_then(|env| std::env::var(env).ok())
-        .filter(|value| !value.is_empty()))
-}
-
-fn configure_dashboard_child_environment(
-    command: &mut tokio::process::Command,
-    provider_key: Option<(&str, &str)>,
-) {
-    configure_dashboard_child_environment_from(command, std::env::vars_os(), provider_key);
-}
-
-fn configure_dashboard_child_environment_from(
-    command: &mut tokio::process::Command,
-    parent_environment: impl IntoIterator<Item = (OsString, OsString)>,
-    provider_key: Option<(&str, &str)>,
-) {
-    command.env_clear();
-    for (name, value) in parent_environment {
-        if dashboard_child_environment_variable_allowed(&name) {
-            command.env(name, value);
-        }
-    }
-
-    if let Some((name, value)) = provider_key {
-        command.env(name, value);
-    }
-}
-
-/// Variables every platform must forward to a spawned job.
-///
-/// Unlike poppler or ffmpeg, this child is BookForge itself doing the actual
-/// provider calls, so withholding network configuration does not harden
-/// anything — it just breaks the run. `reqwest` reads the proxy variables
-/// (in both cases) and some distributions rely on the `SSL_CERT_*` pair to
-/// locate a trust store; drop them and a dashboard-launched job fails to
-/// reach the provider on exactly the machines where the CLI works, which is
-/// a miserable thing to debug. `RUST_LOG` is kept so a job launched from the
-/// dashboard can be traced like any other.
-const DASHBOARD_CHILD_NETWORK_VARIABLES: &[&str] = &[
-    "HTTP_PROXY",
-    "HTTPS_PROXY",
-    "ALL_PROXY",
-    "NO_PROXY",
-    "SSL_CERT_FILE",
-    "SSL_CERT_DIR",
-    "RUST_LOG",
-];
-
-fn dashboard_child_environment_variable_allowed(name: &std::ffi::OsStr) -> bool {
-    let name = name.to_string_lossy();
-
-    // Proxy variables are conventionally read in either case, so match both.
-    if DASHBOARD_CHILD_NETWORK_VARIABLES
-        .iter()
-        .any(|allowed| name.eq_ignore_ascii_case(allowed))
-    {
-        return true;
-    }
-
-    // Exactly one of the following blocks compiles, so each is written as the
-    // function's trailing expression rather than an early return.
-    #[cfg(windows)]
-    {
-        [
-            "PATH",
-            "SYSTEMROOT",
-            "TEMP",
-            "TMP",
-            "USERPROFILE",
-            "APPDATA",
-            "LOCALAPPDATA",
-        ]
-        .iter()
-        .any(|allowed| name.eq_ignore_ascii_case(allowed))
-    }
-
-    #[cfg(unix)]
-    {
-        matches!(
-            name.as_ref(),
-            "PATH" | "HOME" | "LANG" | "LANGUAGE" | "TMPDIR"
-        ) || name.starts_with("LC_")
-    }
-
-    #[cfg(not(any(windows, unix)))]
-    {
-        name == "PATH"
-    }
-}
-
 fn lock_keys(state: &AppState) -> Result<MutexGuard<'_, HashMap<String, String>>> {
     state
         .keys
@@ -2859,157 +412,11 @@ fn lock_keys(state: &AppState) -> Result<MutexGuard<'_, HashMap<String, String>>
         .map_err(|_| anyhow::anyhow!("dashboard API key store is unavailable"))
 }
 
-/// Lock serializing manual corrections for one job.
-///
-/// Applying a correction is a read-modify-write across the whole book, so two
-/// concurrent requests can both read the same pre-correction snapshot and the
-/// later rename publishes an EPUB missing the earlier edit. The registry is
-/// keyed by job id, so corrections to different books still run in parallel.
-fn job_correction_lock(state: &AppState, job_id: &str) -> Result<Arc<tokio::sync::Mutex<()>>> {
-    let mut locks = state
-        .correction_locks
-        .lock()
-        .map_err(|_| anyhow::anyhow!("correction lock registry is unavailable"))?;
-    Ok(locks.entry(job_id.to_string()).or_default().clone())
-}
-
-fn dashboard_options_payload() -> DashboardOptions {
-    DashboardOptions {
-        languages: LANGUAGE_OPTIONS,
-        providers: vec![
-            ProviderOption {
-                id: "mock",
-                label: "mock (offline test)",
-                models: MOCK_MODELS,
-                default_model: "mock-identity",
-                requires_base_url: false,
-                requires_key: false,
-            },
-            ProviderOption {
-                id: "deepseek",
-                label: "deepseek",
-                models: DEEPSEEK_MODELS,
-                default_model: "deepseek-v4-flash",
-                requires_base_url: false,
-                requires_key: true,
-            },
-            ProviderOption {
-                id: "openrouter",
-                label: "openrouter",
-                models: OPENROUTER_MODELS,
-                default_model: "openrouter/auto",
-                requires_base_url: false,
-                requires_key: true,
-            },
-            ProviderOption {
-                id: "openai-compatible",
-                label: "openai-compatible",
-                models: OPENAI_COMPATIBLE_MODELS,
-                default_model: "gpt-4o-mini",
-                requires_base_url: true,
-                requires_key: true,
-            },
-        ],
-        audio_providers: vec![
-            AudioProviderOption {
-                id: "mock",
-                label: "mock (offline test)",
-                models: AUDIO_MOCK_MODELS,
-                default_model: "mock-silence",
-                default_voice: "mock",
-                formats: &["wav"],
-                default_format: "wav",
-                requires_voice: false,
-                requires_key: false,
-                supports_auto_model: false,
-                supports_instructions: false,
-                supports_speed: true,
-                max_chars: 40_000,
-                model_max_chars: BTreeMap::new(),
-            },
-            AudioProviderOption {
-                id: "openai",
-                label: "OpenAI-compatible",
-                models: AUDIO_OPENAI_MODELS,
-                default_model: "gpt-4o-mini-tts",
-                default_voice: "alloy",
-                formats: &["mp3", "opus", "aac", "flac", "wav", "pcm"],
-                default_format: "mp3",
-                requires_voice: false,
-                requires_key: true,
-                supports_auto_model: false,
-                supports_instructions: true,
-                supports_speed: true,
-                max_chars: 4_096,
-                model_max_chars: BTreeMap::new(),
-            },
-            AudioProviderOption {
-                id: "gemini",
-                label: "Gemini TTS",
-                models: AUDIO_GEMINI_MODELS,
-                default_model: "gemini-3.1-flash-tts-preview",
-                default_voice: "Kore",
-                formats: &["wav", "pcm"],
-                default_format: "wav",
-                requires_voice: false,
-                requires_key: true,
-                supports_auto_model: false,
-                supports_instructions: true,
-                supports_speed: false,
-                max_chars: 4_096,
-                model_max_chars: BTreeMap::new(),
-            },
-            AudioProviderOption {
-                id: "elevenlabs",
-                label: "ElevenLabs",
-                models: AUDIO_ELEVENLABS_MODELS,
-                default_model: "",
-                default_voice: "",
-                formats: &["mp3", "opus", "wav", "pcm"],
-                default_format: "mp3",
-                requires_voice: true,
-                requires_key: true,
-                supports_auto_model: true,
-                supports_instructions: false,
-                supports_speed: true,
-                max_chars: bookforge_audio::elevenlabs_model_max_input_chars(
-                    "eleven_multilingual_v2",
-                ),
-                model_max_chars: AUDIO_ELEVENLABS_MODELS
-                    .iter()
-                    .map(|model| {
-                        (
-                            *model,
-                            bookforge_audio::elevenlabs_model_max_input_chars(model),
-                        )
-                    })
-                    .collect(),
-            },
-        ],
-        ffmpeg_available: bookforge_audio::ffmpeg_available(),
-    }
-}
-
 fn field_value(fields: &HashMap<String, String>, key: &str) -> Option<String> {
     fields
         .get(key)
         .map(|value| value.trim().to_string())
         .filter(|value| !value.is_empty())
-}
-
-fn audio_provider_max_chars(provider: &str, model: &str) -> usize {
-    match provider {
-        "mock" => 40_000,
-        "elevenlabs" if model.is_empty() => 10_000,
-        "elevenlabs" => bookforge_audio::elevenlabs_model_max_input_chars(model),
-        "openai" | "gemini" => 4_096,
-        _ => 4_096,
-    }
-}
-
-fn truthy_field(fields: &HashMap<String, String>, key: &str) -> bool {
-    field_value(fields, key)
-        .is_some_and(|value| matches!(value.as_str(), "true" | "on" | "1" | "yes"))
 }
 
 fn strip_epub_suffix(name: &str) -> &str {
@@ -3046,410 +453,6 @@ fn bad_request(message: &str) -> Response {
 fn forbidden(message: &str) -> Response {
     (StatusCode::FORBIDDEN, Json(json!({ "error": message }))).into_response()
 }
-
-fn reject_mutation(headers: &HeaderMap, state: &AppState) -> Option<Response> {
-    if is_cross_site_browser_request(headers) {
-        return Some(forbidden("cross-site dashboard request rejected"));
-    }
-
-    match headers
-        .get(CSRF_HEADER)
-        .and_then(|value| value.to_str().ok())
-    {
-        Some(token) if constant_time_eq(token.as_bytes(), state.csrf_token.as_bytes()) => None,
-        _ => Some(forbidden("missing or invalid dashboard token")),
-    }
-}
-
-fn constant_time_eq(left: &[u8], right: &[u8]) -> bool {
-    if left.len() != right.len() {
-        return false;
-    }
-
-    let mut diff = 0u8;
-    for (left, right) in left.iter().zip(right.iter()) {
-        diff |= left ^ right;
-    }
-    diff == 0
-}
-
-fn is_cross_site_browser_request(headers: &HeaderMap) -> bool {
-    headers
-        .get("sec-fetch-site")
-        .and_then(|value| value.to_str().ok())
-        .is_some_and(|site| {
-            site.eq_ignore_ascii_case("cross-site")
-                || site.eq_ignore_ascii_case("same-site")
-                || site.eq_ignore_ascii_case("none") && has_nonlocal_origin(headers)
-        })
-}
-
-fn has_nonlocal_origin(headers: &HeaderMap) -> bool {
-    headers
-        .get("origin")
-        .and_then(|value| value.to_str().ok())
-        .is_some_and(|origin| {
-            !(origin.starts_with("http://127.0.0.1:")
-                || origin.starts_with("http://localhost:")
-                || origin.starts_with("http://[::1]:"))
-        })
-}
-
-fn generate_csrf_token() -> Result<String> {
-    let mut bytes = [0u8; 16];
-    getrandom::fill(&mut bytes).context("failed to generate dashboard token")?;
-    Ok(hex_bytes(&bytes))
-}
-
-fn hex_bytes(bytes: &[u8]) -> String {
-    let mut out = String::with_capacity(bytes.len() * 2);
-    for byte in bytes {
-        use std::fmt::Write as _;
-        write!(&mut out, "{byte:02x}").expect("writing to string cannot fail");
-    }
-    out
-}
-
-/// Resolve a job's event-log path off the async runtime (sqlite is blocking).
-async fn resolve_events_path(id: String, store_path: PathBuf) -> PathBuf {
-    let fallback = PathBuf::from(format!(".bookforge/runs/{id}/events.jsonl"));
-    let lookup = id.clone();
-    tokio::task::spawn_blocking(move || {
-        let job = JobStore::open(store_path)
-            .ok()
-            .and_then(|store| store.get_job(&lookup).ok().flatten());
-        events_path_for(job.as_ref(), &lookup)
-    })
-    .await
-    .unwrap_or(fallback)
-}
-
-// ---------------------------------------------------------------------------
-// Wire types
-// ---------------------------------------------------------------------------
-
-#[derive(Serialize)]
-struct JobListItem {
-    id: String,
-    status: String,
-    provider: String,
-    model: String,
-    target_lang: String,
-    input_path: String,
-    total_segments: usize,
-    done: usize,
-    succeeded: usize,
-    cached: usize,
-    needs_review: usize,
-    failed: usize,
-    retry_pending: usize,
-    input_tokens: u64,
-    output_tokens: u64,
-}
-
-impl JobListItem {
-    fn new(job: &JobRecord, summary: &JobSummary) -> Self {
-        let done = summary.succeeded + summary.cached + summary.needs_review + summary.failed;
-        Self {
-            id: job.id.clone(),
-            status: summary.status.clone(),
-            provider: job.provider.clone(),
-            model: job.model.clone(),
-            target_lang: job.target_lang.clone(),
-            input_path: job.input_path.display().to_string(),
-            total_segments: summary.total_segments,
-            done,
-            succeeded: summary.succeeded,
-            cached: summary.cached,
-            needs_review: summary.needs_review,
-            failed: summary.failed,
-            retry_pending: summary.retry_pending,
-            input_tokens: summary.input_tokens,
-            output_tokens: summary.output_tokens,
-        }
-    }
-}
-
-#[derive(Serialize)]
-struct JobDetail {
-    id: String,
-    input_path: String,
-    output_path: String,
-    provider: String,
-    model: String,
-    source_lang: Option<String>,
-    target_lang: String,
-    status: String,
-    state: RunState,
-}
-
-#[derive(Debug, Serialize)]
-struct RuntimeMutableSettings {
-    batch_max_output_tokens: Option<u32>,
-    batch_max_items: usize,
-    batch_target_tokens: usize,
-    concurrency: usize,
-    qa: crate::QaMode,
-    double_check: bookforge_core::DoubleCheckMode,
-    validate_output: bool,
-    provider_max_attempts: usize,
-    adaptive_concurrency: bool,
-    adaptive_batch_sizing: bool,
-}
-
-#[derive(Debug, Serialize)]
-struct RuntimeIdentity {
-    provider: String,
-    model: String,
-    source_language: Option<String>,
-    target_language: String,
-    profile: String,
-    prompt_version: String,
-}
-
-#[derive(Debug, Serialize)]
-struct RuntimeLeaseView {
-    state: &'static str,
-    pid: Option<u32>,
-    instance_id: Option<String>,
-    heartbeat_at_ms: Option<u64>,
-    last_loaded_revision: Option<u64>,
-    last_applied_revision: Option<u64>,
-    error: Option<String>,
-}
-
-#[derive(Debug, Serialize)]
-struct RuntimeSettingsView {
-    effective: RuntimeMutableSettings,
-    overrides: super::reconfigure::RunConfigOverrides,
-    revision: u64,
-    applied_revision: u64,
-    changed_fields: Vec<String>,
-    next_boundary: Vec<String>,
-    application_state: &'static str,
-    live: bool,
-    editable: bool,
-    resumable_work: bool,
-    lease: RuntimeLeaseView,
-    identity: RuntimeIdentity,
-}
-
-fn runtime_settings_view(
-    store_path: &std::path::Path,
-    id: &str,
-    lease_stale_after: Duration,
-) -> Result<Option<RuntimeSettingsView>> {
-    let store = JobStore::open(store_path)?;
-    let Some(job) = store.get_job(id)? else {
-        return Ok(None);
-    };
-    let Some(snapshot) = store.load_job_config_snapshot(id)? else {
-        return Ok(None);
-    };
-    let mut settings = snapshot.settings.to_settings();
-    let loaded = super::reconfigure::load_overrides_document_for_job(id)?;
-    let (revision, overrides) = loaded
-        .map(|loaded| (loaded.revision, loaded.overrides))
-        .unwrap_or_default();
-    super::reconfigure::apply_overrides_to_settings(&mut settings, &overrides);
-    let qa = overrides
-        .qa
-        .unwrap_or_else(|| crate::QaMode::from_snapshot(&snapshot.qa_mode));
-    let validate_output = overrides
-        .validate_output
-        .unwrap_or(snapshot.validate_output);
-    let changed_fields = overrides.changed_fields();
-    let next_boundary = overrides.application_boundaries();
-    let lease_state = crate::control::runtime_lease_state(id, lease_stale_after);
-    let (lease, live, applied_revision) = match lease_state {
-        crate::control::RuntimeLeaseState::Fresh(lease) => (
-            RuntimeLeaseView {
-                state: "fresh",
-                pid: Some(lease.pid),
-                instance_id: Some(lease.instance_id.clone()),
-                heartbeat_at_ms: Some(lease.heartbeat_at_ms),
-                last_loaded_revision: Some(lease.last_loaded_revision),
-                last_applied_revision: Some(lease.last_applied_revision),
-                error: None,
-            },
-            true,
-            lease.last_applied_revision,
-        ),
-        crate::control::RuntimeLeaseState::Stale(lease) => (
-            RuntimeLeaseView {
-                state: "stale",
-                pid: Some(lease.pid),
-                instance_id: Some(lease.instance_id.clone()),
-                heartbeat_at_ms: Some(lease.heartbeat_at_ms),
-                last_loaded_revision: Some(lease.last_loaded_revision),
-                last_applied_revision: Some(lease.last_applied_revision),
-                error: None,
-            },
-            false,
-            lease.last_applied_revision,
-        ),
-        crate::control::RuntimeLeaseState::Missing => (
-            RuntimeLeaseView {
-                state: "missing",
-                pid: None,
-                instance_id: None,
-                heartbeat_at_ms: None,
-                last_loaded_revision: None,
-                last_applied_revision: None,
-                error: None,
-            },
-            false,
-            0,
-        ),
-        crate::control::RuntimeLeaseState::Invalid(error) => (
-            RuntimeLeaseView {
-                state: "invalid",
-                pid: None,
-                instance_id: None,
-                heartbeat_at_ms: None,
-                last_loaded_revision: None,
-                last_applied_revision: None,
-                error: Some(error),
-            },
-            false,
-            0,
-        ),
-    };
-    // Translation is only one part of the resumable pipeline. A stopped,
-    // paused, or orphaned-running job may have no pending segments while QA,
-    // double-check, rebuild, validation, or reporting still remains.
-    let resumable_work = !store.resumable_segment_ids(id)?.is_empty()
-        || job_status_has_unfinished_pipeline_work(&job.status);
-    let editable = job_status_has_unfinished_pipeline_work(&job.status) && resumable_work;
-    let application_state = if !live {
-        "resume_required"
-    } else if revision > applied_revision {
-        "next_boundary"
-    } else {
-        "live"
-    };
-    Ok(Some(RuntimeSettingsView {
-        effective: RuntimeMutableSettings {
-            batch_max_output_tokens: settings.provider.batch_max_output_tokens,
-            batch_max_items: settings.batch.max_items,
-            batch_target_tokens: settings.batch.target_tokens,
-            concurrency: settings.scheduler.concurrency,
-            qa,
-            double_check: settings.double_check.mode,
-            validate_output,
-            provider_max_attempts: settings.provider.provider_max_attempts,
-            adaptive_concurrency: settings.adaptive_concurrency,
-            adaptive_batch_sizing: settings.batch.adaptive_sizing,
-        },
-        overrides,
-        revision,
-        applied_revision,
-        changed_fields,
-        next_boundary,
-        application_state,
-        live,
-        editable,
-        resumable_work,
-        lease,
-        identity: RuntimeIdentity {
-            provider: snapshot.provider,
-            model: snapshot.model,
-            source_language: snapshot.source_language,
-            target_language: snapshot.target_language,
-            profile: format!("{:?}", snapshot.profile),
-            prompt_version: snapshot.prompt_version,
-        },
-    }))
-}
-
-#[derive(Serialize)]
-struct DashboardOptions {
-    languages: &'static [&'static str],
-    providers: Vec<ProviderOption>,
-    audio_providers: Vec<AudioProviderOption>,
-    ffmpeg_available: bool,
-}
-
-#[derive(Serialize)]
-struct ProviderOption {
-    id: &'static str,
-    label: &'static str,
-    models: &'static [&'static str],
-    default_model: &'static str,
-    requires_base_url: bool,
-    requires_key: bool,
-}
-
-#[derive(Serialize)]
-struct AudioProviderOption {
-    id: &'static str,
-    label: &'static str,
-    models: &'static [&'static str],
-    default_model: &'static str,
-    default_voice: &'static str,
-    formats: &'static [&'static str],
-    default_format: &'static str,
-    requires_voice: bool,
-    requires_key: bool,
-    supports_auto_model: bool,
-    supports_instructions: bool,
-    supports_speed: bool,
-    max_chars: usize,
-    /// Per-model input ceilings, so the browser never keeps its own copy of a
-    /// provider's limits. Empty when every model shares `max_chars`.
-    model_max_chars: BTreeMap<&'static str, usize>,
-}
-
-impl JobDetail {
-    fn new(id: String, job: Option<JobRecord>, state: RunState) -> Self {
-        let provider = job
-            .as_ref()
-            .map(|j| j.provider.clone())
-            .or_else(|| state.provider.clone())
-            .unwrap_or_default();
-        let model = job
-            .as_ref()
-            .map(|j| j.model.clone())
-            .or_else(|| state.model.clone())
-            .unwrap_or_default();
-        let input_path = job
-            .as_ref()
-            .map(|j| j.input_path.display().to_string())
-            .or_else(|| state.input_path.clone())
-            .unwrap_or_default();
-        let output_path = job
-            .as_ref()
-            .map(|j| j.output_path.display().to_string())
-            .or_else(|| state.output_path.clone())
-            .unwrap_or_default();
-        Self {
-            id,
-            input_path,
-            output_path,
-            provider,
-            model,
-            source_lang: job.as_ref().and_then(|j| j.source_lang.clone()),
-            target_lang: job
-                .as_ref()
-                .map(|j| j.target_lang.clone())
-                .unwrap_or_default(),
-            status: job.as_ref().map(|j| j.status.clone()).unwrap_or_else(|| {
-                if state.finished {
-                    "done".into()
-                } else if state.paused {
-                    "paused".into()
-                } else {
-                    "running".into()
-                }
-            }),
-            state,
-        }
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Error plumbing
-// ---------------------------------------------------------------------------
 
 /// Wraps any error so handlers can use `?`; renders as a 500 JSON body.
 struct AppError(anyhow::Error);
@@ -3495,2113 +498,5 @@ fn open_in_browser(url: &str) -> Result<()> {
     Ok(())
 }
 
-// ---------------------------------------------------------------------------
-// Inline frontend (no build step — vanilla JS, mirrors the review.rs pattern)
-// ---------------------------------------------------------------------------
-
-const DASHBOARD_HTML_TEMPLATE: &str = include_str!("serve/dashboard.html");
-const DASHBOARD_CSS: &str = include_str!("serve/dashboard.css");
-const DASHBOARD_JS: &str = include_str!("serve/dashboard.js");
-
-fn assemble_dashboard_html(template: &str, css: &str, js: &str) -> String {
-    let template = template.replace("\r\n", "\n");
-    let css = css.replace("\r\n", "\n");
-    let js = js.replace("\r\n", "\n");
-
-    template
-        .replace("{{BOOKFORGE_DASHBOARD_CSS}}", &css)
-        .replace("{{BOOKFORGE_DASHBOARD_JS}}", &js)
-}
-
-static DASHBOARD_HTML: std::sync::LazyLock<String> = std::sync::LazyLock::new(|| {
-    assemble_dashboard_html(DASHBOARD_HTML_TEMPLATE, DASHBOARD_CSS, DASHBOARD_JS)
-});
-
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use axum::http::HeaderValue;
-
-    const TEST_HOST: &str = "127.0.0.1:8765";
-    const TEST_DEADLOCK_TIMEOUT: Duration = Duration::from_secs(30);
-
-    fn test_state(token: &str) -> AppState {
-        AppState {
-            refresh: Duration::from_millis(250),
-            csrf_token: token.to_string(),
-            host_port: 8765,
-            upload_dir: PathBuf::from(UPLOAD_DIR),
-            keys: Arc::new(Mutex::new(HashMap::new())),
-            elevenlabs_voices: Arc::new(Mutex::new(None)),
-            audio_cancels: Arc::new(Mutex::new(HashMap::new())),
-            store_path: default_store_path(),
-            runtime_lease_stale_after: crate::control::RUNTIME_LEASE_STALE_AFTER,
-            correction_locks: Arc::new(Mutex::new(HashMap::new())),
-            resume_launches: None,
-            resume_child_environments: None,
-            audio_restart_cancels: None,
-        }
-    }
-
-    /// Like [`test_state`], but pointed at an isolated store path instead of
-    /// the process-relative default — lets tests exercise the store-backed
-    /// mutation endpoints against a temp-dir database without chdir'ing the
-    /// (shared, per-process) current directory, which would race across
-    /// parallel test threads.
-    fn test_state_with_store(token: &str, store_path: PathBuf) -> AppState {
-        AppState {
-            store_path,
-            ..test_state(token)
-        }
-    }
-
-    fn test_state_with_upload_dir(token: &str, upload_dir: PathBuf) -> AppState {
-        AppState {
-            upload_dir,
-            ..test_state(token)
-        }
-    }
-
-    #[test]
-    fn bind_must_be_loopback() {
-        let local: SocketAddr = "127.0.0.1:8765".parse().unwrap();
-        let remote: SocketAddr = "0.0.0.0:8765".parse().unwrap();
-
-        assert!(require_loopback_bind(local).is_ok());
-        assert!(require_loopback_bind(remote).is_err());
-    }
-
-    #[test]
-    fn dashboard_host_header_allows_only_loopback_names_on_bound_port() {
-        for host in [
-            "127.0.0.1:8765",
-            "localhost:8765",
-            "LOCALHOST:8765",
-            "[::1]:8765",
-        ] {
-            assert!(
-                dashboard_host_value_allowed(host, 8765),
-                "{host} should be allowed"
-            );
-        }
-
-        for host in [
-            "",
-            "127.0.0.1",
-            "127.0.0.1:8766",
-            " 127.0.0.1:8765",
-            "127.0.0.1:8765 ",
-            "127.0.0.1.evil.test:8765",
-            "localhost.evil.test:8765",
-            "evil.test:8765",
-            "127.0.0.1:8765.evil.test",
-            "127.0.0.1:08765",
-            "[::1]:8766",
-            "[::ffff:127.0.0.1]:8765",
-        ] {
-            assert!(
-                !dashboard_host_value_allowed(host, 8765),
-                "{host} should be rejected"
-            );
-        }
-    }
-
-    #[test]
-    fn provider_key_envs_cover_supported_cloud_providers() {
-        assert_eq!(provider_key_env("deepseek"), Some("DEEPSEEK_API_KEY"));
-        assert_eq!(provider_key_env("openrouter"), Some("OPENROUTER_API_KEY"));
-        assert_eq!(
-            provider_key_env("openai-compatible"),
-            Some("OPENAI_API_KEY")
-        );
-        assert!(provider_key_env("mock").is_none());
-    }
-
-    #[test]
-    fn dashboard_child_environment_omits_unneeded_provider_secrets() {
-        let mut command = tokio::process::Command::new("bookforge");
-        let fake_parent_environment = [
-            (
-                OsString::from("DEEPSEEK_API_KEY"),
-                OsString::from("fake-parent-secret"),
-            ),
-            (OsString::from("PATH"), OsString::from("fake-path")),
-        ];
-        configure_dashboard_child_environment_from(
-            &mut command,
-            fake_parent_environment,
-            Some(("ELEVENLABS_API_KEY", "required-child-key")),
-        );
-
-        let child_environment = command
-            .as_std()
-            .get_envs()
-            .map(|(name, value)| (name.to_os_string(), value.map(ToOwned::to_owned)))
-            .collect::<HashMap<_, _>>();
-        assert!(!child_environment.contains_key(std::ffi::OsStr::new("DEEPSEEK_API_KEY")));
-        assert_eq!(
-            child_environment
-                .get(std::ffi::OsStr::new("PATH"))
-                .and_then(|value| value.as_deref()),
-            Some(std::ffi::OsStr::new("fake-path"))
-        );
-        assert_eq!(
-            child_environment
-                .get(std::ffi::OsStr::new("ELEVENLABS_API_KEY"))
-                .and_then(|value| value.as_deref()),
-            Some(std::ffi::OsStr::new("required-child-key"))
-        );
-    }
-
-    #[test]
-    fn dashboard_options_include_common_languages_and_models() {
-        let options = dashboard_options_payload();
-        assert!(options.languages.contains(&"Italian"));
-        assert!(options.languages.contains(&"English"));
-        assert!(options.languages.contains(&"Toki Pona"));
-
-        let deepseek = options
-            .providers
-            .iter()
-            .find(|provider| provider.id == "deepseek")
-            .expect("deepseek provider option should exist");
-        assert_eq!(deepseek.default_model, "deepseek-v4-flash");
-        assert!(deepseek.models.contains(&"deepseek-v4-pro"));
-
-        let openrouter = options
-            .providers
-            .iter()
-            .find(|provider| provider.id == "openrouter")
-            .expect("openrouter provider option should exist");
-        assert!(openrouter.models.contains(&"google/gemini-2.5-flash"));
-
-        let gemini_tts = options
-            .audio_providers
-            .iter()
-            .find(|provider| provider.id == "gemini")
-            .expect("Gemini TTS provider option should exist");
-        assert_eq!(gemini_tts.default_voice, "Kore");
-        assert_eq!(gemini_tts.formats, &["wav", "pcm"]);
-        assert!(!gemini_tts.supports_speed);
-
-        let elevenlabs = options
-            .audio_providers
-            .iter()
-            .find(|provider| provider.id == "elevenlabs")
-            .expect("ElevenLabs provider option should exist");
-        assert!(elevenlabs.requires_voice);
-        assert!(elevenlabs.supports_auto_model);
-        assert!(!elevenlabs.supports_instructions);
-        assert_eq!(elevenlabs.models.first(), Some(&"eleven_v3"));
-        assert_eq!(elevenlabs.default_model, "");
-        assert_eq!(elevenlabs.max_chars, 10_000);
-        assert_eq!(audio_provider_max_chars("openai", "anything"), 4_096);
-        assert_eq!(audio_provider_max_chars("elevenlabs", ""), 10_000);
-        assert_eq!(audio_provider_max_chars("elevenlabs", "eleven_v3"), 5_000);
-        assert_eq!(
-            audio_provider_max_chars("elevenlabs", "eleven_flash_v2_5"),
-            40_000
-        );
-    }
-
-    /// The browser used to carry its own table of ElevenLabs per-model limits,
-    /// which is a copy that can drift from the one synthesis actually enforces.
-    /// The options payload now serves them, so guard both halves: the payload
-    /// agrees with the canonical function, and the JS no longer hardcodes them.
-    #[test]
-    fn audio_options_serve_per_model_limits_so_the_browser_keeps_no_copy() {
-        let options = dashboard_options_payload();
-        let elevenlabs = options
-            .audio_providers
-            .iter()
-            .find(|provider| provider.id == "elevenlabs")
-            .expect("elevenlabs provider should be offered");
-
-        assert!(
-            !elevenlabs.model_max_chars.is_empty(),
-            "per-model limits must reach the browser"
-        );
-        for model in elevenlabs.models {
-            assert_eq!(
-                elevenlabs.model_max_chars.get(model).copied(),
-                Some(bookforge_audio::elevenlabs_model_max_input_chars(model)),
-                "served limit for {model} must match the synthesis path"
-            );
-        }
-
-        assert!(
-            DASHBOARD_JS.contains("provider.model_max_chars"),
-            "the browser must read limits from the payload"
-        );
-        for stale in ["eleven_flash_v2_5:40000", "eleven_v3:5000"] {
-            assert!(
-                !DASHBOARD_JS.contains(stale),
-                "dashboard.js must not reintroduce a hardcoded limit table ({stale})"
-            );
-        }
-    }
-
-    #[test]
-    fn dashboard_escapes_dynamic_html_fields() {
-        assert!(DASHBOARD_HTML.contains("function esc(value)"));
-        assert!(DASHBOARD_HTML.contains("${esc(d.id)}"));
-        assert!(DASHBOARD_HTML.contains("${esc(body)}"));
-        assert!(DASHBOARD_HTML.contains("${esc(a.id)}"));
-        assert!(DASHBOARD_HTML.contains("${esc(w.sourcePath)}"));
-        assert!(DASHBOARD_HTML.contains("${esc(audioWarningMessage(warning))}"));
-    }
-
-    #[test]
-    fn dashboard_openai_compatible_base_url_must_be_https() {
-        assert!(dashboard_base_url_uses_https("https://api.example.com/v1"));
-        assert!(!dashboard_base_url_uses_https("http://api.example.com/v1"));
-        assert!(!dashboard_base_url_uses_https("https://"));
-        assert!(!dashboard_base_url_uses_https("not a url"));
-    }
-
-    #[tokio::test]
-    async fn child_startup_check_reports_immediate_success_exit() -> Result<()> {
-        let mut child = tokio::process::Command::new(std::env::current_exe()?)
-            .arg("--help")
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .spawn()?;
-
-        let expected = tokio::time::timeout(TEST_DEADLOCK_TIMEOUT, child.wait())
-            .await
-            .expect("help child should exit before the deadlock guard")?;
-        let status = child_exit_status_after(&mut child, Duration::ZERO)
-            .await?
-            .expect("completed help child should have an exit status");
-
-        assert_eq!(status, expected);
-        assert!(expected.success());
-        Ok(())
-    }
-
-    #[test]
-    fn mutating_routes_require_dashboard_token() {
-        let state = test_state("token-123");
-        let headers = HeaderMap::new();
-        assert!(reject_mutation(&headers, &state).is_some());
-
-        let mut headers = HeaderMap::new();
-        headers.insert(CSRF_HEADER, HeaderValue::from_static("token-123"));
-        assert!(reject_mutation(&headers, &state).is_none());
-    }
-
-    #[test]
-    fn mutating_routes_reject_cross_site_browser_requests() {
-        let state = test_state("token-123");
-        let mut headers = HeaderMap::new();
-        headers.insert(CSRF_HEADER, HeaderValue::from_static("token-123"));
-        headers.insert("sec-fetch-site", HeaderValue::from_static("cross-site"));
-
-        assert!(reject_mutation(&headers, &state).is_some());
-    }
-
-    #[test]
-    fn csrf_token_compare_matches_only_exact_token() {
-        assert!(constant_time_eq(b"token-123", b"token-123"));
-        assert!(!constant_time_eq(b"token-123", b"token-124"));
-        assert!(!constant_time_eq(b"token-123", b"token-1234"));
-    }
-
-    #[tokio::test]
-    async fn dashboard_rejects_untrusted_host_header_before_serving_html() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .uri("/")
-                    .header("host", "evil.test:8765")
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    }
-
-    #[tokio::test]
-    async fn dashboard_serves_loopback_host_header() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .uri("/")
-                    .header("host", "localhost:8765")
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(
-            response.headers().get("content-security-policy"),
-            Some(&HeaderValue::from_static(DASHBOARD_CONTENT_SECURITY_POLICY))
-        );
-        for (name, value) in [
-            ("x-frame-options", "DENY"),
-            ("x-content-type-options", "nosniff"),
-            ("referrer-policy", "no-referrer"),
-            ("cache-control", "no-store"),
-        ] {
-            assert_eq!(
-                response.headers().get(name),
-                Some(&HeaderValue::from_static(value))
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn retry_endpoint_rejects_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/jobs/not-real/retry")
-                    .header("host", TEST_HOST)
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    }
-
-    #[tokio::test]
-    async fn audiobook_endpoint_rejects_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let body =
-            "--B\r\nContent-Disposition: form-data; name=\"provider\"\r\n\r\nmock\r\n--B--\r\n";
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/audiobook")
-                    .header("host", TEST_HOST)
-                    .header("content-type", "multipart/form-data; boundary=B")
-                    .body(Body::from(body))
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    }
-
-    #[tokio::test]
-    async fn audiobook_estimate_endpoint_rejects_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let body =
-            "--B\r\nContent-Disposition: form-data; name=\"provider\"\r\n\r\nmock\r\n--B--\r\n";
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/audiobook/estimate")
-                    .header("host", TEST_HOST)
-                    .header("content-type", "multipart/form-data; boundary=B")
-                    .body(Body::from(body))
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    }
-
-    #[tokio::test]
-    async fn elevenlabs_voices_without_key_returns_conflict() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .uri("/api/audio/voices?provider=elevenlabs")
-                    .header("host", TEST_HOST)
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::CONFLICT);
-    }
-
-    #[test]
-    fn resolved_model_uses_last_synthesis_id_colon() {
-        assert_eq!(
-            resolved_model_from_synthesis_id(
-                "elevenlabs:https://api.elevenlabs.io:443/v1:eleven_v3"
-            ),
-            Some("eleven_v3")
-        );
-    }
-
-    #[test]
-    fn audiobook_command_args_omit_auto_model_and_include_explicit_model() {
-        let build = |model| {
-            audiobook_command_args(
-                Path::new("book.epub"),
-                Path::new("audio-out"),
-                "elevenlabs",
-                model,
-                "voice-id",
-                "mp3",
-                1.0,
-                2_000,
-                4,
-                None,
-                None,
-                true,
-                true,
-                Some("ELEVENLABS_API_KEY"),
-                &AudiobookCommandOptions::default(),
-            )
-        };
-
-        let auto = build(None);
-        assert!(!auto.iter().any(|arg| arg == "--model"));
-
-        let explicit = build(Some("eleven_flash_v2_5"));
-        let model_flag = explicit
-            .iter()
-            .position(|arg| arg == "--model")
-            .expect("explicit model should add --model");
-        assert_eq!(
-            explicit.get(model_flag + 1),
-            Some(&OsString::from("eleven_flash_v2_5"))
-        );
-    }
-
-    #[test]
-    fn audiobook_command_args_include_only_advanced_flags_that_are_set() {
-        let build = |advanced: &AudiobookCommandOptions| {
-            audiobook_command_args(
-                Path::new("book.epub"),
-                Path::new("audio-out"),
-                "elevenlabs",
-                Some("eleven_flash_v2_5"),
-                "voice-id",
-                "mp3",
-                1.0,
-                2_000,
-                4,
-                None,
-                None,
-                true,
-                true,
-                Some("ELEVENLABS_API_KEY"),
-                advanced,
-            )
-        };
-
-        let defaults = build(&AudiobookCommandOptions::default());
-        for flag in ["--single", "--loudnorm", "--seed", "--language"] {
-            assert!(!defaults.iter().any(|arg| arg == flag), "included {flag}");
-        }
-
-        let configured = build(&AudiobookCommandOptions {
-            single: true,
-            loudnorm: true,
-            seed: Some(u32::MAX),
-            language: Some("pt-BR".to_string()),
-            ..AudiobookCommandOptions::default()
-        });
-        for flag in ["--single", "--loudnorm", "--seed", "--language"] {
-            assert!(configured.iter().any(|arg| arg == flag), "omitted {flag}");
-        }
-        assert!(configured.iter().any(|arg| arg == "4294967295"));
-        assert!(configured.iter().any(|arg| arg == "pt-BR"));
-    }
-
-    #[test]
-    fn audiobook_language_validation_matches_dashboard_contract() {
-        for valid in ["it", "en-US", "pt-BR"] {
-            assert!(valid_audio_language(valid), "rejected {valid:?}");
-        }
-        for invalid in ["../etc", "toolongtoken", ""] {
-            assert!(!valid_audio_language(invalid), "accepted {invalid:?}");
-        }
-    }
-
-    #[test]
-    fn audiobook_gap_values_are_clamped_to_ten_seconds() {
-        assert_eq!(clamp_audio_gap(0), 0);
-        assert_eq!(clamp_audio_gap(600), 600);
-        assert_eq!(clamp_audio_gap(99_999), 10_000);
-    }
-
-    #[tokio::test]
-    async fn audiobook_cancel_rejects_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/audiobooks/example/cancel")
-                    .header("host", TEST_HOST)
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    }
-
-    fn write_test_audiobook_operation(
-        upload_dir: &Path,
-        id: &str,
-        manifest: serde_json::Value,
-        process: serde_json::Value,
-    ) -> PathBuf {
-        let out_dir = upload_dir.join(format!("audiobook-{id}"));
-        std::fs::create_dir_all(&out_dir).expect("operation directory should be created");
-        std::fs::write(
-            out_dir.join("manifest.json"),
-            serde_json::to_vec(&manifest).expect("manifest should serialize"),
-        )
-        .expect("manifest should be written");
-        std::fs::write(
-            out_dir.join("process.json"),
-            serde_json::to_vec(&process).expect("process state should serialize"),
-        )
-        .expect("process state should be written");
-        out_dir
-    }
-
-    #[tokio::test]
-    async fn audiobook_index_scans_durable_operations_newest_first() {
-        let temp = tempfile::tempdir().expect("temp dir should be created");
-        write_test_audiobook_operation(
-            temp.path(),
-            "older",
-            json!({
-                "title": "Older Book",
-                "status": "running",
-                "completed_chunks": 1,
-                "chapters": 1,
-                "chunks": [{"status": "synthesized"}],
-                "updated_at_ms": 10,
-            }),
-            json!({"status": "running", "updated_at_ms": 11}),
-        );
-        write_test_audiobook_operation(
-            temp.path(),
-            "newer",
-            json!({
-                "title": "Newer Book",
-                "status": "succeeded",
-                "completed_chunks": 2,
-                "chapters": 1,
-                "chunks": [{"status": "synthesized"}, {"status": "cached"}],
-                "updated_at_ms": 20,
-            }),
-            json!({
-                "status": "succeeded",
-                "warnings": [{"message": "chapter markers were unavailable"}],
-                "updated_at_ms": 21,
-            }),
-        );
-        std::fs::create_dir_all(temp.path().join("not-an-audiobook"))
-            .expect("unrelated directory should be created");
-
-        let router = dashboard_router(test_state_with_upload_dir(
-            "token-123",
-            temp.path().to_path_buf(),
-        ));
-        let response = get_route(&router, "/api/audiobooks").await;
-        assert_eq!(response.status(), StatusCode::OK);
-        let payload = response_json(response).await;
-        let items = payload.as_array().expect("index should be an array");
-        assert_eq!(items.len(), 2);
-        assert_eq!(items[0]["id"], "newer");
-        assert_eq!(items[0]["title"], "Newer Book");
-        assert_eq!(items[0]["total_chunks"], 2);
-        assert_eq!(
-            items[0]["warnings"][0]["message"],
-            "chapter markers were unavailable"
-        );
-        assert_eq!(items[1]["id"], "older");
-    }
-
-    #[tokio::test]
-    async fn audiobook_cancel_uses_persisted_pid_after_server_restart() {
-        let temp = tempfile::tempdir().expect("temp dir should be created");
-        let out_dir = write_test_audiobook_operation(
-            temp.path(),
-            "restartable",
-            json!({"status": "running", "chunks": []}),
-            json!({
-                "status": "running",
-                "pid": 4242,
-                "auto_model": true,
-                "warnings": ["stitch fallback"],
-                "updated_at_ms": 10,
-            }),
-        );
-        let cancelled = Arc::new(Mutex::new(Vec::new()));
-        let mut state = test_state_with_upload_dir("token-123", temp.path().to_path_buf());
-        state.audio_restart_cancels = Some(cancelled.clone());
-        assert!(state.audio_cancels.lock().unwrap().is_empty());
-        let router = dashboard_router(state);
-
-        let response = post_json(
-            &router,
-            "/api/audiobooks/restartable/cancel",
-            Some("token-123"),
-            json!({}),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(*cancelled.lock().unwrap(), vec![4242]);
-        let process: serde_json::Value = serde_json::from_slice(
-            &std::fs::read(out_dir.join("process.json")).expect("process state should remain"),
-        )
-        .expect("process state should be JSON");
-        assert_eq!(process["status"], "cancelled");
-        assert_eq!(process["auto_model"], true);
-        assert_eq!(process["warnings"][0], "stitch fallback");
-    }
-
-    #[tokio::test]
-    async fn audiobook_source_resolves_finished_translation_by_job_id() {
-        let fixture = build_mutation_fixture();
-        build_fixture_epub(&fixture.output_path);
-        let state = test_state_with_store("token-123", fixture.store_path.clone());
-        let fields = HashMap::from([("source_job_id".to_string(), fixture.job_id.clone())]);
-
-        let source = resolve_audiobook_source(
-            &state,
-            &fields,
-            None,
-            "ignored-upload-name.epub".to_string(),
-        )
-        .await
-        .expect("finished translation output should resolve");
-
-        assert_eq!(source.file_name, "output.epub");
-        assert!(source.bytes.starts_with(b"PK"));
-    }
-
-    #[tokio::test]
-    async fn audiobook_artifact_supports_ranges_and_rejects_unsatisfiable_ranges() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let temp = tempfile::tempdir().expect("temp dir should be created");
-        let out_dir = temp.path().join("audiobook-range-test");
-        std::fs::create_dir_all(&out_dir).expect("operation directory should be created");
-        std::fs::write(out_dir.join("audiobook.m4b"), b"0123456789")
-            .expect("artifact should be written");
-        let router = dashboard_router(test_state_with_upload_dir(
-            "token-123",
-            temp.path().to_path_buf(),
-        ));
-
-        let partial = router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .uri("/api/audiobooks/range-test/artifact")
-                    .header("host", TEST_HOST)
-                    .header("range", "bytes=2-5")
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-        assert_eq!(partial.status(), StatusCode::PARTIAL_CONTENT);
-        assert_eq!(
-            partial.headers().get("content-range"),
-            Some(&HeaderValue::from_static("bytes 2-5/10"))
-        );
-        assert_eq!(
-            partial.headers().get("accept-ranges"),
-            Some(&HeaderValue::from_static("bytes"))
-        );
-        let body = axum::body::to_bytes(partial.into_body(), usize::MAX)
-            .await
-            .expect("partial body should read");
-        assert_eq!(&body[..], b"2345");
-
-        let unsatisfiable = router
-            .oneshot(
-                Request::builder()
-                    .uri("/api/audiobooks/range-test/artifact")
-                    .header("host", TEST_HOST)
-                    .header("range", "bytes=20-")
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-        assert_eq!(unsatisfiable.status(), StatusCode::RANGE_NOT_SATISFIABLE);
-        assert_eq!(
-            unsatisfiable.headers().get("content-range"),
-            Some(&HeaderValue::from_static("bytes */10"))
-        );
-    }
-
-    #[test]
-    fn audiobook_operation_ids_cannot_escape_the_upload_directory() {
-        assert!(valid_audiobook_id("1234-safe_id"));
-        for invalid in ["", "../escape", "with/slash", "with.dot", "with space"] {
-            assert!(!valid_audiobook_id(invalid), "accepted {invalid:?}");
-        }
-    }
-
-    #[tokio::test]
-    async fn control_endpoints_reject_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let router = dashboard_router(test_state("token-123"));
-        for command in ["pause", "resume", "stop"] {
-            let response = router
-                .clone()
-                .oneshot(
-                    Request::builder()
-                        .method("POST")
-                        .uri(format!("/api/jobs/not-real/{command}"))
-                        .header("host", TEST_HOST)
-                        .body(Body::empty())
-                        .expect("request should build"),
-                )
-                .await
-                .expect("route should respond");
-
-            assert_eq!(response.status(), StatusCode::FORBIDDEN, "{command}");
-        }
-    }
-
-    #[tokio::test]
-    async fn estimate_endpoint_rejects_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        // A well-formed multipart body so the Multipart extractor succeeds and
-        // the handler's own CSRF check is what rejects the request.
-        let body =
-            "--B\r\nContent-Disposition: form-data; name=\"provider\"\r\n\r\nmock\r\n--B--\r\n";
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/estimate")
-                    .header("host", TEST_HOST)
-                    .header("content-type", "multipart/form-data; boundary=B")
-                    .body(Body::from(body))
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::FORBIDDEN);
-    }
-
-    #[tokio::test]
-    async fn glossary_mutations_reject_missing_dashboard_token() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let add = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/glossary")
-                    .header("host", TEST_HOST)
-                    .header("content-type", "application/json")
-                    .body(Body::from(
-                        r#"{"source":"a","target":"b","source_language":"English","target_language":"Italian"}"#,
-                    ))
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-        assert_eq!(add.status(), StatusCode::FORBIDDEN);
-
-        let remove = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("DELETE")
-                    .uri("/api/glossary/1")
-                    .header("host", TEST_HOST)
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-        assert_eq!(remove.status(), StatusCode::FORBIDDEN);
-    }
-
-    #[tokio::test]
-    async fn translate_rejects_non_https_openai_compatible_base_url() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let body = concat!(
-            "--B\r\n",
-            "Content-Disposition: form-data; name=\"file\"; filename=\"book.epub\"\r\n",
-            "Content-Type: application/epub+zip\r\n\r\n",
-            "not-a-real-epub\r\n",
-            "--B\r\n",
-            "Content-Disposition: form-data; name=\"target\"\r\n\r\n",
-            "Italian\r\n",
-            "--B\r\n",
-            "Content-Disposition: form-data; name=\"provider\"\r\n\r\n",
-            "openai-compatible\r\n",
-            "--B\r\n",
-            "Content-Disposition: form-data; name=\"base_url\"\r\n\r\n",
-            "http://api.example.test/v1\r\n",
-            "--B\r\n",
-            "Content-Disposition: form-data; name=\"model\"\r\n\r\n",
-            "test-model\r\n",
-            "--B\r\n",
-            "Content-Disposition: form-data; name=\"api_key\"\r\n\r\n",
-            "sk-test-key\r\n",
-            "--B--\r\n",
-        );
-
-        let response = dashboard_router(test_state("token-123"))
-            .oneshot(
-                Request::builder()
-                    .method("POST")
-                    .uri("/api/translate")
-                    .header("host", TEST_HOST)
-                    .header(CSRF_HEADER, "token-123")
-                    .header("content-type", "multipart/form-data; boundary=B")
-                    .body(Body::from(body))
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-
-        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-    }
-
-    #[test]
-    fn glossary_scope_and_category_parse_with_sane_defaults() {
-        assert_eq!(parse_glossary_scope("book"), GlossaryScopeKind::Book);
-        assert_eq!(parse_glossary_scope("series"), GlossaryScopeKind::Series);
-        assert_eq!(parse_glossary_scope("whatever"), GlossaryScopeKind::Global);
-        assert_eq!(parse_glossary_category("place"), GlossaryCategory::Place);
-        assert_eq!(parse_glossary_category("mystery"), GlossaryCategory::Other);
-    }
-
-    #[test]
-    fn dashboard_ships_all_screen_renderers() {
-        for marker in [
-            "function renderLibrary",
-            "function renderWizard",
-            "function renderAudiobook",
-            "function renderProgress",
-            "function drawReview",
-            "function drawValidation",
-            "function renderGlossary",
-            "function bfAudioEstimate",
-            "/api/audiobook/estimate",
-            "Advanced",
-        ] {
-            assert!(DASHBOARD_HTML.contains(marker), "missing {marker}");
-        }
-        assert!(!DASHBOARD_HTML.contains("fonts.googleapis.com"));
-        assert!(!DASHBOARD_HTML.contains("fonts.gstatic.com"));
-    }
-
-    #[test]
-    fn dashboard_assets_reassemble_byte_stably() {
-        use sha2::{Digest, Sha256};
-
-        assert_eq!(DASHBOARD_HTML.len(), 113_292);
-        assert!(!DASHBOARD_HTML.contains("{{BOOKFORGE_DASHBOARD_CSS}}"));
-        assert!(!DASHBOARD_HTML.contains("{{BOOKFORGE_DASHBOARD_JS}}"));
-        let digest = Sha256::digest(DASHBOARD_HTML.as_bytes());
-        let digest_hex: String = digest.iter().map(|byte| format!("{byte:02x}")).collect();
-        assert_eq!(
-            digest_hex,
-            "908c441831ab32699345eedd9157909c26c8477d7c1e893d5c176daeb351c486"
-        );
-
-        let crlf = |asset: &str| asset.replace("\r\n", "\n").replace('\n', "\r\n");
-        assert_eq!(
-            assemble_dashboard_html(
-                &crlf(DASHBOARD_HTML_TEMPLATE),
-                &crlf(DASHBOARD_CSS),
-                &crlf(DASHBOARD_JS),
-            ),
-            *DASHBOARD_HTML,
-        );
-    }
-
-    #[test]
-    fn dashboard_ships_runtime_editor_and_inline_retry_guidance() {
-        for marker in [
-            "function drawRuntimeSettings",
-            "function bfSaveRuntimeSettings",
-            "RuntimeConfigChanged",
-            "function bfReviewRetrySubmit",
-            "Stop the job before queuing a retry",
-            "function bfReviewStopForRetry",
-            "function showResumeKeyEntry",
-            "function bfResumeWithKey",
-        ] {
-            assert!(DASHBOARD_HTML.contains(marker), "missing {marker}");
-        }
-        assert!(
-            !DASHBOARD_HTML.contains("window.prompt"),
-            "retry guidance must use the inline editor"
-        );
-    }
-
-    #[test]
-    fn dashboard_posts_include_csrf_header() {
-        assert!(DASHBOARD_HTML.contains(CSRF_TOKEN_PLACEHOLDER));
-        assert!(DASHBOARD_HTML.contains("headers: { [CSRF_HEADER]: CSRF_TOKEN }"));
-    }
-
-    #[test]
-    fn csrf_token_is_hex_encoded_random_bytes() {
-        let token = generate_csrf_token().expect("token should generate");
-        assert_eq!(token.len(), 32);
-        assert!(token.chars().all(|ch| ch.is_ascii_hexdigit()));
-    }
-
-    // -----------------------------------------------------------------------
-    // Segment mutation endpoints: CSRF rejection + isolated-store end-to-end.
-    //
-    // These build a real completed job (real EPUB parsed with `read_epub`,
-    // real `Segment`s from `build_segments`, real store rows) in a per-test
-    // temp directory so `save_manual_translation` / `set_segment_flag` /
-    // `retry_segment_with_guidance` exercise the same code paths production
-    // traffic does. Every test gets its own `tempfile::TempDir` and its own
-    // `AppState::store_path` (via `test_state_with_store`), so nothing here
-    // touches the process's current directory and tests are parallel-safe.
-    // -----------------------------------------------------------------------
-
-    use bookforge_core::segment::BlockTranslation;
-    use bookforge_store::{CreateJob, SaveTranslation};
-    use std::io::Write as _;
-
-    const FIXTURE_CONTAINER_XML: &str = r#"<?xml version="1.0" encoding="UTF-8"?>
-<container version="1.0" xmlns="urn:oasis:names:tc:opendocument:xmlns:container">
-  <rootfiles>
-    <rootfile full-path="content.opf" media-type="application/oebps-package+xml"/>
-  </rootfiles>
-</container>"#;
-
-    const FIXTURE_OPF: &str = r#"<?xml version="1.0" encoding="UTF-8"?>
-<package xmlns="http://www.idpf.org/2007/opf" version="3.0" unique-identifier="uid">
-  <metadata xmlns:dc="http://purl.org/dc/elements/1.1/">
-    <dc:identifier id="uid">serve-fixture</dc:identifier>
-    <dc:title>Serve Fixture</dc:title>
-    <dc:language>en</dc:language>
-  </metadata>
-  <manifest>
-    <item id="ch1" href="chapter1.xhtml" media-type="application/xhtml+xml"/>
-    <item id="ch2" href="chapter2.xhtml" media-type="application/xhtml+xml"/>
-  </manifest>
-  <spine>
-    <itemref idref="ch1"/>
-    <itemref idref="ch2"/>
-  </spine>
-</package>"#;
-
-    const FIXTURE_CHAPTER_ONE: &str = r#"<?xml version="1.0" encoding="UTF-8"?>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head><title>Fixture Chapter One</title></head>
-<body>
-<p>The lantern flickered in the old library.</p>
-</body>
-</html>"#;
-
-    const FIXTURE_CHAPTER_TWO: &str = r#"<?xml version="1.0" encoding="UTF-8"?>
-<html xmlns="http://www.w3.org/1999/xhtml">
-<head><title>Fixture Chapter Two</title></head>
-<body>
-<p>Rain tapped steadily against the windowpane.</p>
-</body>
-</html>"#;
-
-    fn build_fixture_epub(path: &std::path::Path) {
-        use zip::{CompressionMethod, ZipWriter, write::SimpleFileOptions};
-
-        let file = std::fs::File::create(path).expect("fixture EPUB should be creatable");
-        let mut zip = ZipWriter::new(file);
-        let stored = SimpleFileOptions::default().compression_method(CompressionMethod::Stored);
-        let deflated = SimpleFileOptions::default().compression_method(CompressionMethod::Deflated);
-
-        zip.start_file("mimetype", stored).unwrap();
-        zip.write_all(b"application/epub+zip").unwrap();
-        zip.start_file("META-INF/container.xml", deflated).unwrap();
-        zip.write_all(FIXTURE_CONTAINER_XML.as_bytes()).unwrap();
-        zip.start_file("content.opf", deflated).unwrap();
-        zip.write_all(FIXTURE_OPF.as_bytes()).unwrap();
-        zip.start_file("chapter1.xhtml", deflated).unwrap();
-        zip.write_all(FIXTURE_CHAPTER_ONE.as_bytes()).unwrap();
-        zip.start_file("chapter2.xhtml", deflated).unwrap();
-        zip.write_all(FIXTURE_CHAPTER_TWO.as_bytes()).unwrap();
-        zip.finish().unwrap();
-    }
-
-    /// A completed two-segment job (one segment per chapter, one block per
-    /// segment) backed by an isolated temp-dir store and a real rebuildable
-    /// output path — everything the mutation endpoints under test touch.
-    struct MutationFixture {
-        // Held only to keep the temp directory alive for the fixture's
-        // lifetime; never read directly.
-        _temp: tempfile::TempDir,
-        store_path: PathBuf,
-        output_path: PathBuf,
-        job_id: String,
-        segment_a: String,
-        segment_b: String,
-        csrf: String,
-    }
-
-    fn build_mutation_fixture() -> MutationFixture {
-        build_mutation_fixture_for_provider("mock", "mock-identity", None)
-    }
-
-    fn build_mutation_fixture_for_provider(
-        provider: &str,
-        model: &str,
-        api_key_env: Option<&str>,
-    ) -> MutationFixture {
-        let temp = tempfile::tempdir().expect("temp dir should be created");
-        let input_path = temp.path().join("input.epub");
-        build_fixture_epub(&input_path);
-        let output_path = temp.path().join("output.epub");
-        let store_path = temp.path().join("jobs.sqlite");
-
-        let store = JobStore::open(&store_path).expect("store should open");
-        let job = store
-            .create_job(CreateJob {
-                input: &input_path,
-                output: &output_path,
-                source_lang: Some("English"),
-                target_lang: "Italian",
-                provider,
-                model,
-                base_url: None,
-                api_key_env,
-                book_id: None,
-                series_id: None,
-            })
-            .expect("job should be created");
-
-        let book = bookforge_epub::read_epub(&input_path).expect("fixture EPUB should parse");
-        let settings = bookforge_core::TranslationProfile::Balanced.resolve();
-        let segments = bookforge_core::segment::build_segments(&book, &settings.segmentation)
-            .expect("segments should build");
-        // `read_epub`/`build_segments` also synthesize a segment for the OPF
-        // `dc:title` metadata (its own section, ahead of the spine chapters),
-        // so a two-chapter book yields three segments total. Only the two
-        // chapter segments are used as `segment_a`/`segment_b` below; the
-        // metadata segment is still inserted and translated like any other
-        // so the fixture matches what a real job actually persists.
-        let chapter_segments = segments
-            .iter()
-            .filter(|segment| segment.section_id.0 != "sec_metadata_opf")
-            .collect::<Vec<_>>();
-        assert_eq!(
-            chapter_segments.len(),
-            2,
-            "fixture EPUB (one paragraph per chapter) should yield exactly two chapter segments"
-        );
-
-        store
-            .insert_segments(&job.id, &segments, "v1", provider, model, "test_ns")
-            .expect("segments should insert");
-
-        for segment in &segments {
-            let blocks = segment
-                .source
-                .blocks
-                .iter()
-                .map(|block| BlockTranslation {
-                    block_id: block.block_id.clone(),
-                    text: format!("[IT] {}", block.text),
-                })
-                .collect::<Vec<_>>();
-            let translated_text = blocks
-                .iter()
-                .map(|block| block.text.as_str())
-                .collect::<Vec<_>>()
-                .join("\n\n");
-            store
-                .save_translation(SaveTranslation {
-                    job_id: &job.id,
-                    segment_id: &segment.id.0,
-                    translated_text: &translated_text,
-                    blocks: &blocks,
-                    input_tokens: Some(10),
-                    input_cached_tokens: Some(0),
-                    output_tokens: Some(12),
-                    tokens_estimated: false,
-                    provider,
-                    model,
-                    prompt_version: "v1",
-                })
-                .expect("translation should save");
-        }
-        store
-            .mark_job_complete(&job.id)
-            .expect("job should complete");
-
-        let snapshot = bookforge_core::RunConfigSnapshot {
-            input_path: input_path.clone(),
-            input_snapshot_path: Some(input_path.clone()),
-            input_sha256: Some("test-sha".to_string()),
-            output_path: output_path.clone(),
-            events_path: None,
-            report_json_path: None,
-            report_markdown_path: None,
-            source_language: Some("English".to_string()),
-            target_language: "Italian".to_string(),
-            creator: None,
-            provider: provider.to_string(),
-            model: model.to_string(),
-            base_url: None,
-            api_key_env: api_key_env.map(str::to_string),
-            profile: settings.profile,
-            provider_preset: None,
-            prompt_version: "v1".to_string(),
-            cache_namespace: "test_ns".to_string(),
-            book_id: None,
-            series_id: None,
-            glossary_budget_tokens: 800,
-            glossary_format: bookforge_core::GlossaryFormat::Json,
-            prompt_extra: None,
-            glossary_fingerprint: String::new(),
-            glossary_terms: Vec::new(),
-            context_window: 0,
-            context_budget_tokens: 1200,
-            context_scope: bookforge_core::config::ContextScope::Chapter,
-            style_fingerprint: String::new(),
-            style_rendered_block: String::new(),
-            entities_fingerprint: String::new(),
-            entities_rendered_block: String::new(),
-            bilingual_mode: bookforge_core::BilingualMode::Replace,
-            bilingual_separator: " / ".to_string(),
-            bilingual_style: bookforge_core::BilingualStyle::Minimal,
-            bilingual_css: None,
-            fallback: None,
-            finalize: bookforge_core::run_snapshot::FinalizeCheckpointSnapshot::default(),
-            qa_mode: "off".to_string(),
-            validate_output: false,
-            settings: bookforge_core::ResolvedRunSettingsSnapshot::from_settings(&settings),
-        };
-        store
-            .update_job_config_snapshot(&job.id, &snapshot)
-            .expect("snapshot should persist");
-
-        MutationFixture {
-            store_path,
-            output_path,
-            job_id: job.id,
-            segment_a: chapter_segments[0].id.0.clone(),
-            segment_b: chapter_segments[1].id.0.clone(),
-            csrf: "fixture-csrf-token".to_string(),
-            _temp: temp,
-        }
-    }
-
-    /// Sends `body` to `uri` with the given CSRF header value (or none), and
-    /// returns the response.
-    async fn post_json(
-        router: &Router,
-        uri: &str,
-        csrf: Option<&str>,
-        body: serde_json::Value,
-    ) -> Response {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let mut builder = Request::builder()
-            .method("POST")
-            .uri(uri)
-            .header("host", TEST_HOST)
-            .header("content-type", "application/json");
-        if let Some(token) = csrf {
-            builder = builder.header(CSRF_HEADER, token);
-        }
-        router
-            .clone()
-            .oneshot(
-                builder
-                    .body(Body::from(body.to_string()))
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond")
-    }
-
-    async fn get_route(router: &Router, uri: &str) -> Response {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .uri(uri)
-                    .header("host", TEST_HOST)
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond")
-    }
-
-    async fn response_json(response: Response) -> serde_json::Value {
-        let bytes = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .expect("response body should read");
-        serde_json::from_slice(&bytes).expect("response should be JSON")
-    }
-
-    fn make_stopped_fixture_resumable(fixture: &MutationFixture) {
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        store
-            .request_segment_retry(&fixture.job_id, &fixture.segment_b, None)
-            .expect("completed segment should become retry-pending");
-        store
-            .mark_job_stopped(&fixture.job_id)
-            .expect("job should become stopped");
-    }
-
-    fn clean_runtime_files(job_id: &str) {
-        let _ = std::fs::remove_dir_all(bookforge_core::run_dir_for_job(job_id));
-    }
-
-    #[tokio::test]
-    async fn dashboard_reconfigure_is_typed_revisioned_and_csrf_protected() {
-        let fixture = build_mutation_fixture();
-        make_stopped_fixture_resumable(&fixture);
-        clean_runtime_files(&fixture.job_id);
-        let router = dashboard_router(test_state_with_store(
-            &fixture.csrf,
-            fixture.store_path.clone(),
-        ));
-        let uri = format!("/api/jobs/{}/reconfigure", fixture.job_id);
-
-        let initial = get_route(&router, &uri).await;
-        assert_eq!(initial.status(), StatusCode::OK);
-        let initial = response_json(initial).await;
-        assert_eq!(initial["revision"], 0);
-        assert_eq!(initial["applied_revision"], 0);
-        assert_eq!(initial["application_state"], "resume_required");
-        assert_eq!(initial["lease"]["state"], "missing");
-        assert_eq!(initial["identity"]["provider"], "mock");
-        assert_eq!(initial["identity"]["model"], "mock-identity");
-        assert_eq!(initial["editable"], true);
-
-        let body = json!({ "concurrency": 2 });
-        let missing = post_json(&router, &uri, None, body.clone()).await;
-        assert_eq!(missing.status(), StatusCode::FORBIDDEN);
-        let wrong = post_json(&router, &uri, Some("wrong-token"), body).await;
-        assert_eq!(wrong.status(), StatusCode::FORBIDDEN);
-        assert!(
-            !crate::commands::reconfigure::overrides_path_for_job(&fixture.job_id).exists(),
-            "rejected mutations must not create a sidecar"
-        );
-
-        let unknown = post_json(
-            &router,
-            &uri,
-            Some(&fixture.csrf),
-            json!({ "model": "immutable-model" }),
-        )
-        .await;
-        assert!(unknown.status().is_client_error());
-        assert!(
-            !crate::commands::reconfigure::overrides_path_for_job(&fixture.job_id).exists(),
-            "unknown or immutable fields must not create a sidecar"
-        );
-
-        let first = post_json(
-            &router,
-            &uri,
-            Some(&fixture.csrf),
-            json!({
-                "batch_max_output_tokens": 12000,
-                "batch_max_items": 2,
-                "batch_target_tokens": 3000,
-                "concurrency": 2,
-                "qa": "all",
-                "double_check": "Formatting",
-                "validate_output": true,
-                "provider_max_attempts": 5,
-                "adaptive_concurrency": false,
-                "adaptive_batch_sizing": false
-            }),
-        )
-        .await;
-        assert_eq!(first.status(), StatusCode::OK);
-        let first = response_json(first).await;
-        assert_eq!(first["revision"], 1);
-        assert_eq!(first["effective"]["concurrency"], 2);
-        assert_eq!(first["effective"]["qa"], "all");
-        assert_eq!(first["effective"]["double_check"], "Formatting");
-        assert_eq!(first["effective"]["validate_output"], true);
-        assert_eq!(first["application_state"], "resume_required");
-        let fields = first["changed_fields"]
-            .as_array()
-            .expect("changed fields should be an array");
-        assert!(fields.iter().any(|field| field == "concurrency"));
-        assert!(fields.iter().any(|field| field == "qa"));
-        let boundaries = first["next_boundary"]
-            .as_array()
-            .expect("boundaries should be an array");
-        for boundary in ["next_request", "next_batch", "next_stage"] {
-            assert!(boundaries.iter().any(|value| value == boundary));
-        }
-
-        let second = post_json(
-            &router,
-            &uri,
-            Some(&fixture.csrf),
-            json!({ "concurrency": 3 }),
-        )
-        .await;
-        assert_eq!(second.status(), StatusCode::OK);
-        let second = response_json(second).await;
-        assert_eq!(second["revision"], 2);
-        assert_eq!(second["effective"]["concurrency"], 3);
-        assert_eq!(second["effective"]["batch_max_items"], 2);
-
-        let replayed = response_json(get_route(&router, &uri).await).await;
-        assert_eq!(replayed["revision"], 2);
-        assert_eq!(replayed["effective"]["concurrency"], 3);
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_controls_require_a_fresh_lease_and_signal_one_when_present() {
-        let fixture = build_mutation_fixture();
-        make_stopped_fixture_resumable(&fixture);
-        clean_runtime_files(&fixture.job_id);
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        state.runtime_lease_stale_after = Duration::from_millis(u64::MAX);
-        let router = dashboard_router(state);
-
-        for command in ["pause", "stop"] {
-            let response = post_json(
-                &router,
-                &format!("/api/jobs/{}/{}", fixture.job_id, command),
-                Some(&fixture.csrf),
-                json!({}),
-            )
-            .await;
-            assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-            let payload = response_json(response).await;
-            assert!(
-                payload["error"]
-                    .as_str()
-                    .unwrap_or_default()
-                    .contains("no live worker")
-            );
-        }
-
-        let lease = crate::control::RuntimeLease {
-            schema_version: 1,
-            instance_id: "dashboard-test-worker".to_string(),
-            pid: std::process::id(),
-            process_started_at_ms: bookforge_core::now_ms(),
-            heartbeat_at_ms: bookforge_core::now_ms(),
-            last_loaded_revision: 4,
-            last_applied_revision: 4,
-        };
-        let runtime_path = crate::control::runtime_path_for_job(&fixture.job_id);
-        std::fs::create_dir_all(runtime_path.parent().expect("runtime parent"))
-            .expect("runtime directory should exist");
-        std::fs::write(
-            &runtime_path,
-            serde_json::to_vec_pretty(&lease).expect("lease should serialize"),
-        )
-        .expect("lease should write");
-
-        let view = response_json(
-            get_route(
-                &router,
-                &format!("/api/jobs/{}/reconfigure", fixture.job_id),
-            )
-            .await,
-        )
-        .await;
-        assert_eq!(view["lease"]["state"], "fresh");
-        assert_eq!(view["live"], true);
-        assert_eq!(view["applied_revision"], 4);
-
-        let resume = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({}),
-        )
-        .await;
-        assert_eq!(resume.status(), StatusCode::OK);
-        let resume = response_json(resume).await;
-        assert_eq!(resume["mode"], "signaled");
-        assert_eq!(
-            bookforge_core::read_control_file(&bookforge_core::control_path_for_job(
-                &fixture.job_id
-            ))
-            .expect("control file should read"),
-            bookforge_core::ControlCommand::Resume
-        );
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_resume_uses_remembered_key_in_scrubbed_child_environment() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        const KEY_ENV: &str = "BOOKFORGE_SERVE_TEST_DEEPSEEK_KEY";
-        const SESSION_KEY: &str = "dashboard-session-key";
-
-        let fixture =
-            build_mutation_fixture_for_provider("deepseek", "deepseek-chat", Some(KEY_ENV));
-        make_stopped_fixture_resumable(&fixture);
-        clean_runtime_files(&fixture.job_id);
-
-        let launches = Arc::new(AtomicUsize::new(0));
-        let environments = Arc::new(Mutex::new(Vec::new()));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        lock_keys(&state).expect("key store should lock").extend([
-            ("deepseek".to_string(), SESSION_KEY.to_string()),
-            (
-                "openrouter".to_string(),
-                "unrelated-session-key".to_string(),
-            ),
-        ]);
-        state.resume_launches = Some(launches.clone());
-        state.resume_child_environments = Some(environments.clone());
-        let router = dashboard_router(state);
-
-        let response = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({}),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-        let response = response_json(response).await;
-        assert_eq!(response["mode"], "spawned");
-        assert!(!response.to_string().contains(SESSION_KEY));
-        assert_eq!(launches.load(Ordering::SeqCst), 1);
-
-        let environments = environments.lock().expect("environments should lock");
-        let environment = environments
-            .first()
-            .expect("one resume environment should be captured");
-        assert_eq!(
-            environment
-                .get(std::ffi::OsStr::new(KEY_ENV))
-                .and_then(|value| value.as_deref()),
-            Some(std::ffi::OsStr::new(SESSION_KEY))
-        );
-        for (_, unrelated_env) in PROVIDER_KEY_ENVS {
-            assert!(
-                !environment.contains_key(std::ffi::OsStr::new(unrelated_env)),
-                "{unrelated_env} must not leak into the resume child"
-            );
-        }
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_resume_without_required_key_returns_actionable_error() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        const KEY_ENV: &str = "BOOKFORGE_SERVE_TEST_MISSING_OPENROUTER_KEY_9C71";
-        let fixture =
-            build_mutation_fixture_for_provider("openrouter", "openrouter/auto", Some(KEY_ENV));
-        make_stopped_fixture_resumable(&fixture);
-        clean_runtime_files(&fixture.job_id);
-
-        let launches = Arc::new(AtomicUsize::new(0));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        state.resume_launches = Some(launches.clone());
-        let router = dashboard_router(state);
-
-        let response = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({}),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-        let response = response_json(response).await;
-        assert_eq!(response["requires_api_key"], true);
-        assert_eq!(response["provider"], "openrouter");
-        assert_eq!(response["api_key_env"], KEY_ENV);
-        assert!(
-            response["error"]
-                .as_str()
-                .unwrap_or_default()
-                .contains("supply it to resume")
-        );
-        assert_eq!(launches.load(Ordering::SeqCst), 0);
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_resume_accepts_and_remembers_replacement_key() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        const KEY_ENV: &str = "BOOKFORGE_SERVE_TEST_RESUPPLIED_OPENROUTER_KEY";
-        const REPLACEMENT_KEY: &str = "replacement-dashboard-key";
-
-        let fixture =
-            build_mutation_fixture_for_provider("openrouter", "openrouter/auto", Some(KEY_ENV));
-        make_stopped_fixture_resumable(&fixture);
-        clean_runtime_files(&fixture.job_id);
-
-        let launches = Arc::new(AtomicUsize::new(0));
-        let environments = Arc::new(Mutex::new(Vec::new()));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        let keys = state.keys.clone();
-        state.resume_launches = Some(launches.clone());
-        state.resume_child_environments = Some(environments.clone());
-        let router = dashboard_router(state);
-
-        let response = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({ "api_key": REPLACEMENT_KEY }),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-        assert_eq!(response_json(response).await["mode"], "spawned");
-        assert_eq!(launches.load(Ordering::SeqCst), 1);
-        assert_eq!(
-            keys.lock()
-                .expect("key store should lock")
-                .get("openrouter")
-                .map(String::as_str),
-            Some(REPLACEMENT_KEY)
-        );
-        let environments = environments.lock().expect("environments should lock");
-        let environment = environments
-            .first()
-            .expect("one resume environment should be captured");
-        assert_eq!(
-            environment
-                .get(std::ffi::OsStr::new(KEY_ENV))
-                .and_then(|value| value.as_deref()),
-            Some(std::ffi::OsStr::new(REPLACEMENT_KEY))
-        );
-        assert!(!environment.contains_key(std::ffi::OsStr::new("DEEPSEEK_API_KEY")));
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_missing_worker_resume_launches_exactly_once() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let fixture = build_mutation_fixture();
-        make_stopped_fixture_resumable(&fixture);
-        clean_runtime_files(&fixture.job_id);
-        let launches = Arc::new(AtomicUsize::new(0));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        state.resume_launches = Some(launches.clone());
-        let router = dashboard_router(state);
-        let uri = format!("/api/jobs/{}/resume", fixture.job_id);
-
-        let (first, second) = tokio::join!(
-            post_json(&router, &uri, Some(&fixture.csrf), json!({})),
-            post_json(&router, &uri, Some(&fixture.csrf), json!({}))
-        );
-        assert_eq!(first.status(), StatusCode::OK);
-        assert_eq!(second.status(), StatusCode::OK);
-        let first = response_json(first).await;
-        let second = response_json(second).await;
-        let modes = [
-            first["mode"].as_str().unwrap_or_default(),
-            second["mode"].as_str().unwrap_or_default(),
-        ];
-        assert!(modes.contains(&"spawned"));
-        assert!(modes.contains(&"launching"));
-        assert_eq!(
-            launches.load(Ordering::SeqCst),
-            1,
-            "the atomic launch claim must deduplicate concurrent Resume clicks"
-        );
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_resume_recognizes_finalize_only_work() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let fixture = build_mutation_fixture();
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        assert!(
-            store
-                .resumable_segment_ids(&fixture.job_id)
-                .expect("segment lookup should succeed")
-                .is_empty(),
-            "the fixture should have no translation work left"
-        );
-        store
-            .mark_job_stopped(&fixture.job_id)
-            .expect("job should become stopped during finalization");
-        clean_runtime_files(&fixture.job_id);
-
-        let launches = Arc::new(AtomicUsize::new(0));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        state.resume_launches = Some(launches.clone());
-        let router = dashboard_router(state);
-
-        let view = response_json(
-            get_route(
-                &router,
-                &format!("/api/jobs/{}/reconfigure", fixture.job_id),
-            )
-            .await,
-        )
-        .await;
-        assert_eq!(view["resumable_work"], true);
-        assert_eq!(view["editable"], true);
-
-        let response = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({}),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-        let response = response_json(response).await;
-        assert_eq!(response["mode"], "spawned");
-        assert_eq!(response["forced"], false);
-        assert_eq!(launches.load(Ordering::SeqCst), 1);
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_resume_forces_relaunch_of_a_dead_paused_worker() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let fixture = build_mutation_fixture();
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        store
-            .mark_job_paused(&fixture.job_id)
-            .expect("job should become paused");
-        clean_runtime_files(&fixture.job_id);
-
-        let launches = Arc::new(AtomicUsize::new(0));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        state.resume_launches = Some(launches.clone());
-        let router = dashboard_router(state);
-        let response = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({}),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-        let response = response_json(response).await;
-        assert_eq!(response["mode"], "spawned");
-        assert_eq!(response["forced"], true);
-        assert_eq!(launches.load(Ordering::SeqCst), 1);
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn dashboard_resume_rejects_a_completed_job_without_work() {
-        use std::sync::atomic::{AtomicUsize, Ordering};
-
-        let fixture = build_mutation_fixture();
-        clean_runtime_files(&fixture.job_id);
-        let launches = Arc::new(AtomicUsize::new(0));
-        let mut state = test_state_with_store(&fixture.csrf, fixture.store_path.clone());
-        state.resume_launches = Some(launches.clone());
-        let router = dashboard_router(state);
-
-        let view = response_json(
-            get_route(
-                &router,
-                &format!("/api/jobs/{}/reconfigure", fixture.job_id),
-            )
-            .await,
-        )
-        .await;
-        assert_eq!(view["resumable_work"], false);
-        assert_eq!(view["editable"], false);
-
-        let response = post_json(
-            &router,
-            &format!("/api/jobs/{}/resume", fixture.job_id),
-            Some(&fixture.csrf),
-            json!({}),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-        assert!(
-            response_json(response).await["error"]
-                .as_str()
-                .unwrap_or_default()
-                .contains("no resumable work")
-        );
-        assert_eq!(launches.load(Ordering::SeqCst), 0);
-
-        clean_runtime_files(&fixture.job_id);
-    }
-
-    #[tokio::test]
-    async fn correction_locks_serialize_one_job_and_not_two() {
-        let state = test_state("token-123");
-
-        let a1 = job_correction_lock(&state, "job-a").expect("lock should resolve");
-        let a2 = job_correction_lock(&state, "job-a").expect("lock should resolve");
-        let b = job_correction_lock(&state, "job-b").expect("lock should resolve");
-        assert!(
-            Arc::ptr_eq(&a1, &a2),
-            "the same job must contend on one lock"
-        );
-        assert!(
-            !Arc::ptr_eq(&a1, &b),
-            "different books must not block each other"
-        );
-
-        // Two corrections to the same job must not overlap. Without the lock
-        // both would read the same pre-correction snapshot and the later
-        // rename would publish an EPUB missing the earlier edit.
-        let order = Arc::new(Mutex::new(Vec::<&'static str>::new()));
-        let first = {
-            let lock = a1.clone();
-            let order = order.clone();
-            tokio::spawn(async move {
-                let _guard = lock.lock().await;
-                order.lock().unwrap().push("first-enter");
-                tokio::time::sleep(Duration::from_millis(50)).await;
-                order.lock().unwrap().push("first-exit");
-            })
-        };
-        tokio::time::sleep(Duration::from_millis(10)).await;
-        let second = {
-            let lock = a2.clone();
-            let order = order.clone();
-            tokio::spawn(async move {
-                let _guard = lock.lock().await;
-                order.lock().unwrap().push("second-enter");
-            })
-        };
-        first.await.expect("first correction should finish");
-        second.await.expect("second correction should finish");
-
-        assert_eq!(
-            *order.lock().unwrap(),
-            vec!["first-enter", "first-exit", "second-enter"],
-            "the second correction must wait for the first to finish"
-        );
-    }
-
-    #[tokio::test]
-    async fn save_manual_translation_rejects_missing_or_wrong_csrf_without_mutating_store() {
-        let fixture = build_mutation_fixture();
-        let router = dashboard_router(test_state_with_store(
-            &fixture.csrf,
-            fixture.store_path.clone(),
-        ));
-        let uri = format!(
-            "/api/jobs/{}/segments/{}/translation",
-            fixture.job_id, fixture.segment_a
-        );
-        let body = json!({ "blocks": [{ "block_id": "whatever", "text": "corrupted" }] });
-
-        let missing = post_json(&router, &uri, None, body.clone()).await;
-        assert_eq!(missing.status(), StatusCode::FORBIDDEN);
-
-        let wrong = post_json(&router, &uri, Some("wrong-token"), body).await;
-        assert_eq!(wrong.status(), StatusCode::FORBIDDEN);
-
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        assert!(
-            !store
-                .translation_is_human_corrected(&fixture.job_id, &fixture.segment_a)
-                .expect("lookup should succeed"),
-            "a rejected request must not human-correct the segment"
-        );
-    }
-
-    #[tokio::test]
-    async fn set_segment_flag_rejects_missing_or_wrong_csrf_without_mutating_store() {
-        let fixture = build_mutation_fixture();
-        let router = dashboard_router(test_state_with_store(
-            &fixture.csrf,
-            fixture.store_path.clone(),
-        ));
-        let uri = format!(
-            "/api/jobs/{}/segments/{}/flag",
-            fixture.job_id, fixture.segment_b
-        );
-        let body = json!({ "flagged": true });
-
-        let missing = post_json(&router, &uri, None, body.clone()).await;
-        assert_eq!(missing.status(), StatusCode::FORBIDDEN);
-
-        let wrong = post_json(&router, &uri, Some("wrong-token"), body).await;
-        assert_eq!(wrong.status(), StatusCode::FORBIDDEN);
-
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        let flagged = store
-            .dashboard_flagged_segment_ids(&fixture.job_id)
-            .expect("flags should load");
-        assert!(
-            !flagged.contains(&fixture.segment_b),
-            "a rejected request must not persist a flag"
-        );
-    }
-
-    #[tokio::test]
-    async fn retry_segment_rejects_missing_or_wrong_csrf_without_mutating_store() {
-        let fixture = build_mutation_fixture();
-        let router = dashboard_router(test_state_with_store(
-            &fixture.csrf,
-            fixture.store_path.clone(),
-        ));
-        let uri = format!(
-            "/api/jobs/{}/segments/{}/retry",
-            fixture.job_id, fixture.segment_b
-        );
-        let body = json!({ "guidance": "please redo" });
-
-        let missing = post_json(&router, &uri, None, body.clone()).await;
-        assert_eq!(missing.status(), StatusCode::FORBIDDEN);
-
-        let wrong = post_json(&router, &uri, Some("wrong-token"), body).await;
-        assert_eq!(wrong.status(), StatusCode::FORBIDDEN);
-
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        let guidance = store
-            .load_retry_guidance(&fixture.job_id)
-            .expect("guidance should load");
-        assert!(
-            !guidance.contains_key(&fixture.segment_b),
-            "a rejected request must not persist retry guidance"
-        );
-        let records = store
-            .segment_records(&fixture.job_id)
-            .expect("records should load");
-        let segment_b = records
-            .iter()
-            .find(|record| record.id == fixture.segment_b)
-            .expect("segment_b should exist");
-        assert_eq!(
-            segment_b.status, "succeeded",
-            "a rejected retry must not move the segment to retry_pending"
-        );
-    }
-
-    #[tokio::test]
-    async fn dashboard_review_and_mutation_endpoints_end_to_end() {
-        use axum::{body::Body, http::Request};
-        use tower::ServiceExt;
-
-        let fixture = build_mutation_fixture();
-        let router = dashboard_router(test_state_with_store(
-            &fixture.csrf,
-            fixture.store_path.clone(),
-        ));
-
-        // 1. Review data: per-block source/target text is present for both segments.
-        let review_uri = format!("/api/jobs/{}/review", fixture.job_id);
-        let response = router
-            .clone()
-            .oneshot(
-                Request::builder()
-                    .uri(&review_uri)
-                    .header("host", TEST_HOST)
-                    .body(Body::empty())
-                    .expect("request should build"),
-            )
-            .await
-            .expect("route should respond");
-        assert_eq!(response.status(), StatusCode::OK);
-        let bytes = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .expect("body should read");
-        let review: serde_json::Value =
-            serde_json::from_slice(&bytes).expect("review should be json");
-        let segments = review["segments"].as_array().expect("segments array");
-        // Two chapter segments plus the synthesized OPF `dc:title` metadata
-        // segment (see `build_mutation_fixture`).
-        assert_eq!(segments.len(), 3);
-        for segment in segments {
-            let blocks = segment["blocks"].as_array().expect("blocks array");
-            assert!(!blocks.is_empty(), "each segment should have blocks");
-            for block in blocks {
-                assert!(
-                    !block["target_text"].as_str().unwrap_or_default().is_empty(),
-                    "each block should carry non-empty target text"
-                );
-            }
-        }
-
-        // 2. Save a corrected translation for segment_a.
-        let segment_a_json = segments
-            .iter()
-            .find(|segment| segment["segment_id"] == fixture.segment_a)
-            .expect("segment_a should appear in the review");
-        let block_ids = segment_a_json["blocks"]
-            .as_array()
-            .expect("blocks array")
-            .iter()
-            .map(|block| {
-                block["block_id"]
-                    .as_str()
-                    .expect("block_id should be a string")
-                    .to_string()
-            })
-            .collect::<Vec<_>>();
-        let correction_body = json!({
-            "blocks": block_ids
-                .iter()
-                .map(|id| json!({ "block_id": id, "text": "Corrected by reviewer." }))
-                .collect::<Vec<_>>(),
-        });
-        let translation_uri = format!(
-            "/api/jobs/{}/segments/{}/translation",
-            fixture.job_id, fixture.segment_a
-        );
-        let response = post_json(
-            &router,
-            &translation_uri,
-            Some(&fixture.csrf),
-            correction_body,
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let store = JobStore::open(&fixture.store_path).expect("store should reopen");
-        assert!(
-            store
-                .translation_is_human_corrected(&fixture.job_id, &fixture.segment_a)
-                .expect("lookup should succeed"),
-            "segment_a should be marked human_corrected"
-        );
-        assert!(
-            fixture.output_path.exists(),
-            "the rebuilt output EPUB should exist after the correction"
-        );
-
-        // 3. Flag, then clear, segment_b.
-        let flag_uri = format!(
-            "/api/jobs/{}/segments/{}/flag",
-            fixture.job_id, fixture.segment_b
-        );
-        for flagged in [true, false] {
-            let response = post_json(
-                &router,
-                &flag_uri,
-                Some(&fixture.csrf),
-                json!({ "flagged": flagged }),
-            )
-            .await;
-            assert_eq!(response.status(), StatusCode::OK);
-            let flagged_ids = store
-                .dashboard_flagged_segment_ids(&fixture.job_id)
-                .expect("flags should load");
-            assert_eq!(flagged_ids.contains(&fixture.segment_b), flagged);
-        }
-
-        // 4. Request a retry with guidance for segment_b.
-        let retry_uri = format!(
-            "/api/jobs/{}/segments/{}/retry",
-            fixture.job_id, fixture.segment_b
-        );
-        let response = post_json(
-            &router,
-            &retry_uri,
-            Some(&fixture.csrf),
-            json!({ "guidance": "Please redo more literally." }),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::OK);
-
-        let guidance = store
-            .load_retry_guidance(&fixture.job_id)
-            .expect("guidance should load");
-        assert_eq!(
-            guidance.get(&fixture.segment_b).map(String::as_str),
-            Some("Please redo more literally.")
-        );
-        let records = store
-            .segment_records(&fixture.job_id)
-            .expect("records should load");
-        let segment_b_record = records
-            .iter()
-            .find(|record| record.id == fixture.segment_b)
-            .expect("segment_b should exist");
-        assert_eq!(segment_b_record.status, "retry_pending");
-
-        // 5. A retry request against the now-frozen (human-corrected) segment_a
-        //    is rejected, and does not disturb its correction.
-        let retry_a_uri = format!(
-            "/api/jobs/{}/segments/{}/retry",
-            fixture.job_id, fixture.segment_a
-        );
-        let response = post_json(
-            &router,
-            &retry_a_uri,
-            Some(&fixture.csrf),
-            json!({ "guidance": "try again" }),
-        )
-        .await;
-        assert_eq!(response.status(), StatusCode::BAD_REQUEST);
-        let bytes = axum::body::to_bytes(response.into_body(), usize::MAX)
-            .await
-            .expect("body should read");
-        let payload: serde_json::Value =
-            serde_json::from_slice(&bytes).expect("error body should be json");
-        assert!(
-            payload["error"]
-                .as_str()
-                .unwrap_or_default()
-                .contains("frozen human correction"),
-            "rejection should explain the segment is frozen: {payload}"
-        );
-        assert!(
-            store
-                .translation_is_human_corrected(&fixture.job_id, &fixture.segment_a)
-                .expect("lookup should succeed"),
-            "the rejected retry must not un-freeze the human correction"
-        );
-    }
-}
+mod tests;
